adding column Transition_factor to dataframe expanded_indices

The transition_factor for year j indexes a payment of year j to the niveau of payments in the fixed index_year.

Usage

add_transition_factor(expanded_indices, index_year)

Arguments

expanded_indices

dataframe, see details

index_year

desired year that the payments are indexed to

Value

dataframe indices with new column Transition_factor

Details

expanded_indices must be a dataframe with columns Calendar_year and Index_gross and Index_re.
expanded_indices must contain one row for each historic Calendar_year that shall be used plus 250 future calendar years.

Index_gross is the claim payment development from one year to another, e.g. 0.02 for 2% increase

Index_re is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs. If this column is missing, it will be set to 0.

Examples

expanded_indices <- expand_historic_indices(data.frame(Calendar_year = 2015:2023,
                                                       Index_gross = 0.03),
                                            first_orig_year = 2015,
                                            last_orig_year = 2023,
                                            0.03,
                                            0.02)
add_transition_factor(expanded_indices, index_year = 2022)
#>     Calendar_year Index_gross Index_re Transition_factor
#> 1            2015        0.03     0.00         0.8130915
#> 2            2016        0.03     0.00         0.8374843
#> 3            2017        0.03     0.00         0.8626088
#> 4            2018        0.03     0.00         0.8884870
#> 5            2019        0.03     0.00         0.9151417
#> 6            2020        0.03     0.00         0.9425959
#> 7            2021        0.03     0.00         0.9708738
#> 8            2022        0.03     0.00         1.0000000
#> 9            2023        0.03     0.00         1.0300000
#> 10           2024        0.03     0.02         1.0609000
#> 11           2025        0.03     0.02         1.0927270
#> 12           2026        0.03     0.02         1.1255088
#> 13           2027        0.03     0.02         1.1592741
#> 14           2028        0.03     0.02         1.1940523
#> 15           2029        0.03     0.02         1.2298739
#> 16           2030        0.03     0.02         1.2667701
#> 17           2031        0.03     0.02         1.3047732
#> 18           2032        0.03     0.02         1.3439164
#> 19           2033        0.03     0.02         1.3842339
#> 20           2034        0.03     0.02         1.4257609
#> 21           2035        0.03     0.02         1.4685337
#> 22           2036        0.03     0.02         1.5125897
#> 23           2037        0.03     0.02         1.5579674
#> 24           2038        0.03     0.02         1.6047064
#> 25           2039        0.03     0.02         1.6528476
#> 26           2040        0.03     0.02         1.7024331
#> 27           2041        0.03     0.02         1.7535061
#> 28           2042        0.03     0.02         1.8061112
#> 29           2043        0.03     0.02         1.8602946
#> 30           2044        0.03     0.02         1.9161034
#> 31           2045        0.03     0.02         1.9735865
#> 32           2046        0.03     0.02         2.0327941
#> 33           2047        0.03     0.02         2.0937779
#> 34           2048        0.03     0.02         2.1565913
#> 35           2049        0.03     0.02         2.2212890
#> 36           2050        0.03     0.02         2.2879277
#> 37           2051        0.03     0.02         2.3565655
#> 38           2052        0.03     0.02         2.4272625
#> 39           2053        0.03     0.02         2.5000803
#> 40           2054        0.03     0.02         2.5750828
#> 41           2055        0.03     0.02         2.6523352
#> 42           2056        0.03     0.02         2.7319053
#> 43           2057        0.03     0.02         2.8138625
#> 44           2058        0.03     0.02         2.8982783
#> 45           2059        0.03     0.02         2.9852267
#> 46           2060        0.03     0.02         3.0747835
#> 47           2061        0.03     0.02         3.1670270
#> 48           2062        0.03     0.02         3.2620378
#> 49           2063        0.03     0.02         3.3598989
#> 50           2064        0.03     0.02         3.4606959
#> 51           2065        0.03     0.02         3.5645168
#> 52           2066        0.03     0.02         3.6714523
#> 53           2067        0.03     0.02         3.7815958
#> 54           2068        0.03     0.02         3.8950437
#> 55           2069        0.03     0.02         4.0118950
#> 56           2070        0.03     0.02         4.1322519
#> 57           2071        0.03     0.02         4.2562194
#> 58           2072        0.03     0.02         4.3839060
#> 59           2073        0.03     0.02         4.5154232
#> 60           2074        0.03     0.02         4.6508859
#> 61           2075        0.03     0.02         4.7904125
#> 62           2076        0.03     0.02         4.9341248
#> 63           2077        0.03     0.02         5.0821486
#> 64           2078        0.03     0.02         5.2346130
#> 65           2079        0.03     0.02         5.3916514
#> 66           2080        0.03     0.02         5.5534010
#> 67           2081        0.03     0.02         5.7200030
#> 68           2082        0.03     0.02         5.8916031
#> 69           2083        0.03     0.02         6.0683512
#> 70           2084        0.03     0.02         6.2504017
#> 71           2085        0.03     0.02         6.4379138
#> 72           2086        0.03     0.02         6.6310512
#> 73           2087        0.03     0.02         6.8299827
#> 74           2088        0.03     0.02         7.0348822
#> 75           2089        0.03     0.02         7.2459287
#> 76           2090        0.03     0.02         7.4633065
#> 77           2091        0.03     0.02         7.6872057
#> 78           2092        0.03     0.02         7.9178219
#> 79           2093        0.03     0.02         8.1553566
#> 80           2094        0.03     0.02         8.4000173
#> 81           2095        0.03     0.02         8.6520178
#> 82           2096        0.03     0.02         8.9115783
#> 83           2097        0.03     0.02         9.1789257
#> 84           2098        0.03     0.02         9.4542934
#> 85           2099        0.03     0.02         9.7379222
#> 86           2100        0.03     0.02        10.0300599
#> 87           2101        0.03     0.02        10.3309617
#> 88           2102        0.03     0.02        10.6408906
#> 89           2103        0.03     0.02        10.9601173
#> 90           2104        0.03     0.02        11.2889208
#> 91           2105        0.03     0.02        11.6275884
#> 92           2106        0.03     0.02        11.9764161
#> 93           2107        0.03     0.02        12.3357085
#> 94           2108        0.03     0.02        12.7057798
#> 95           2109        0.03     0.02        13.0869532
#> 96           2110        0.03     0.02        13.4795618
#> 97           2111        0.03     0.02        13.8839487
#> 98           2112        0.03     0.02        14.3004671
#> 99           2113        0.03     0.02        14.7294811
#> 100          2114        0.03     0.02        15.1713656
#> 101          2115        0.03     0.02        15.6265065
#> 102          2116        0.03     0.02        16.0953017
#> 103          2117        0.03     0.02        16.5781608
#> 104          2118        0.03     0.02        17.0755056
#> 105          2119        0.03     0.02        17.5877708
#> 106          2120        0.03     0.02        18.1154039
#> 107          2121        0.03     0.02        18.6588660
#> 108          2122        0.03     0.02        19.2186320
#> 109          2123        0.03     0.02        19.7951909
#> 110          2124        0.03     0.02        20.3890467
#> 111          2125        0.03     0.02        21.0007181
#> 112          2126        0.03     0.02        21.6307396
#> 113          2127        0.03     0.02        22.2796618
#> 114          2128        0.03     0.02        22.9480517
#> 115          2129        0.03     0.02        23.6364932
#> 116          2130        0.03     0.02        24.3455880
#> 117          2131        0.03     0.02        25.0759556
#> 118          2132        0.03     0.02        25.8282343
#> 119          2133        0.03     0.02        26.6030813
#> 120          2134        0.03     0.02        27.4011738
#> 121          2135        0.03     0.02        28.2232090
#> 122          2136        0.03     0.02        29.0699053
#> 123          2137        0.03     0.02        29.9420024
#> 124          2138        0.03     0.02        30.8402625
#> 125          2139        0.03     0.02        31.7654704
#> 126          2140        0.03     0.02        32.7184345
#> 127          2141        0.03     0.02        33.6999875
#> 128          2142        0.03     0.02        34.7109871
#> 129          2143        0.03     0.02        35.7523167
#> 130          2144        0.03     0.02        36.8248863
#> 131          2145        0.03     0.02        37.9296328
#> 132          2146        0.03     0.02        39.0675218
#> 133          2147        0.03     0.02        40.2395475
#> 134          2148        0.03     0.02        41.4467339
#> 135          2149        0.03     0.02        42.6901359
#> 136          2150        0.03     0.02        43.9708400
#> 137          2151        0.03     0.02        45.2899652
#> 138          2152        0.03     0.02        46.6486642
#> 139          2153        0.03     0.02        48.0481241
#> 140          2154        0.03     0.02        49.4895678
#> 141          2155        0.03     0.02        50.9742548
#> 142          2156        0.03     0.02        52.5034825
#> 143          2157        0.03     0.02        54.0785870
#> 144          2158        0.03     0.02        55.7009446
#> 145          2159        0.03     0.02        57.3719729
#> 146          2160        0.03     0.02        59.0931321
#> 147          2161        0.03     0.02        60.8659261
#> 148          2162        0.03     0.02        62.6919038
#> 149          2163        0.03     0.02        64.5726609
#> 150          2164        0.03     0.02        66.5098408
#> 151          2165        0.03     0.02        68.5051360
#> 152          2166        0.03     0.02        70.5602901
#> 153          2167        0.03     0.02        72.6770988
#> 154          2168        0.03     0.02        74.8574117
#> 155          2169        0.03     0.02        77.1031341
#> 156          2170        0.03     0.02        79.4162281
#> 157          2171        0.03     0.02        81.7987150
#> 158          2172        0.03     0.02        84.2526764
#> 159          2173        0.03     0.02        86.7802567
#> 160          2174        0.03     0.02        89.3836644
#> 161          2175        0.03     0.02        92.0651743
#> 162          2176        0.03     0.02        94.8271296
#> 163          2177        0.03     0.02        97.6719435
#> 164          2178        0.03     0.02       100.6021018
#> 165          2179        0.03     0.02       103.6201648
#> 166          2180        0.03     0.02       106.7287698
#> 167          2181        0.03     0.02       109.9306328
#> 168          2182        0.03     0.02       113.2285518
#> 169          2183        0.03     0.02       116.6254084
#> 170          2184        0.03     0.02       120.1241706
#> 171          2185        0.03     0.02       123.7278958
#> 172          2186        0.03     0.02       127.4397326
#> 173          2187        0.03     0.02       131.2629246
#> 174          2188        0.03     0.02       135.2008123
#> 175          2189        0.03     0.02       139.2568367
#> 176          2190        0.03     0.02       143.4345418
#> 177          2191        0.03     0.02       147.7375781
#> 178          2192        0.03     0.02       152.1697054
#> 179          2193        0.03     0.02       156.7347966
#> 180          2194        0.03     0.02       161.4368405
#> 181          2195        0.03     0.02       166.2799457
#> 182          2196        0.03     0.02       171.2683441
#> 183          2197        0.03     0.02       176.4063944
#> 184          2198        0.03     0.02       181.6985862
#> 185          2199        0.03     0.02       187.1495438
#> 186          2200        0.03     0.02       192.7640301
#> 187          2201        0.03     0.02       198.5469510
#> 188          2202        0.03     0.02       204.5033596
#> 189          2203        0.03     0.02       210.6384603
#> 190          2204        0.03     0.02       216.9576141
#> 191          2205        0.03     0.02       223.4663426
#> 192          2206        0.03     0.02       230.1703329
#> 193          2207        0.03     0.02       237.0754428
#> 194          2208        0.03     0.02       244.1877061
#> 195          2209        0.03     0.02       251.5133373
#> 196          2210        0.03     0.02       259.0587374
#> 197          2211        0.03     0.02       266.8304995
#> 198          2212        0.03     0.02       274.8354145
#> 199          2213        0.03     0.02       283.0804770
#> 200          2214        0.03     0.02       291.5728913
#> 201          2215        0.03     0.02       300.3200780
#> 202          2216        0.03     0.02       309.3296804
#> 203          2217        0.03     0.02       318.6095708
#> 204          2218        0.03     0.02       328.1678579
#> 205          2219        0.03     0.02       338.0128936
#> 206          2220        0.03     0.02       348.1532804
#> 207          2221        0.03     0.02       358.5978789
#> 208          2222        0.03     0.02       369.3558152
#> 209          2223        0.03     0.02       380.4364897
#> 210          2224        0.03     0.02       391.8495844
#> 211          2225        0.03     0.02       403.6050719
#> 212          2226        0.03     0.02       415.7132240
#> 213          2227        0.03     0.02       428.1846208
#> 214          2228        0.03     0.02       441.0301594
#> 215          2229        0.03     0.02       454.2610642
#> 216          2230        0.03     0.02       467.8888961
#> 217          2231        0.03     0.02       481.9255630
#> 218          2232        0.03     0.02       496.3833299
#> 219          2233        0.03     0.02       511.2748298
#> 220          2234        0.03     0.02       526.6130747
#> 221          2235        0.03     0.02       542.4114669
#> 222          2236        0.03     0.02       558.6838109
#> 223          2237        0.03     0.02       575.4443252
#> 224          2238        0.03     0.02       592.7076550
#> 225          2239        0.03     0.02       610.4888846
#> 226          2240        0.03     0.02       628.8035512
#> 227          2241        0.03     0.02       647.6676577
#> 228          2242        0.03     0.02       667.0976875
#> 229          2243        0.03     0.02       687.1106181
#> 230          2244        0.03     0.02       707.7239366
#> 231          2245        0.03     0.02       728.9556547
#> 232          2246        0.03     0.02       750.8243244
#> 233          2247        0.03     0.02       773.3490541
#> 234          2248        0.03     0.02       796.5495257
#> 235          2249        0.03     0.02       820.4460115
#> 236          2250        0.03     0.02       845.0593918
#> 237          2251        0.03     0.02       870.4111736
#> 238          2252        0.03     0.02       896.5235088
#> 239          2253        0.03     0.02       923.4192141
#> 240          2254        0.03     0.02       951.1217905
#> 241          2255        0.03     0.02       979.6554442
#> 242          2256        0.03     0.02      1009.0451075
#> 243          2257        0.03     0.02      1039.3164607
#> 244          2258        0.03     0.02      1070.4959546
#> 245          2259        0.03     0.02      1102.6108332
#> 246          2260        0.03     0.02      1135.6891582
#> 247          2261        0.03     0.02      1169.7598329
#> 248          2262        0.03     0.02      1204.8526279
#> 249          2263        0.03     0.02      1240.9982068
#> 250          2264        0.03     0.02      1278.2281530
#> 251          2265        0.03     0.02      1316.5749976
#> 252          2266        0.03     0.02      1356.0722475
#> 253          2267        0.03     0.02      1396.7544149
#> 254          2268        0.03     0.02      1438.6570474
#> 255          2269        0.03     0.02      1481.8167588
#> 256          2270        0.03     0.02      1526.2712615
#> 257          2271        0.03     0.02      1572.0593994
#> 258          2272        0.03     0.02      1619.2211814
#> 259          2273        0.03     0.02      1667.7978168