Cuts out the history part of a matrix with historic and future payments

Cuts out the history part of a matrix with historic and future payments

Usage

cut2history(total_matrix)

Arguments

total_matrix

numeric matrix with n rows and n + 250 columns. Rownames must be the origin years, colnames must be the calendar years

Value

First n columns of total_matrix

Examples

total_matrix <- cbind(skeleton_triangle(2000,2010), skeleton_prediction(2000, 2010))
cut2history(total_matrix)
#>      2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010
#> 2000    0    0    0    0    0    0    0    0    0    0    0
#> 2001    0    0    0    0    0    0    0    0    0    0    0
#> 2002    0    0    0    0    0    0    0    0    0    0    0
#> 2003    0    0    0    0    0    0    0    0    0    0    0
#> 2004    0    0    0    0    0    0    0    0    0    0    0
#> 2005    0    0    0    0    0    0    0    0    0    0    0
#> 2006    0    0    0    0    0    0    0    0    0    0    0
#> 2007    0    0    0    0    0    0    0    0    0    0    0
#> 2008    0    0    0    0    0    0    0    0    0    0    0
#> 2009    0    0    0    0    0    0    0    0    0    0    0
#> 2010    0    0    0    0    0    0    0    0    0    0    0