expand dataframe historic_indices by expected future indices

expand dataframe historic_indices by expected future indices

Usage

expand_historic_indices(
  historic_indices,
  first_orig_year,
  last_orig_year,
  index_gross_future = 0,
  index_re_future = 0
)

Arguments

historic_indices

dataframe, see details

first_orig_year

First origin year with full history.

last_orig_year

Last origin year.

index_gross_future

expected future index for gross claims, default: 0

index_re_future

expected future index for index clause, see details

Value

extended dataframe

Details

historic_indices must be a dataframe with columns Calendar_year and Index_gross and (optional) Index_re.
historic_indices must contain one row for each historic Calendar_year that shall be used

Index_gross is the claim payment development from one year to another, e.g. 0.02 for 2% increase

Index_re is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs. If this column is missing, it will be set to 0.

The function extends the dataframe historic_indices by the column index_re_future if missing and by 250 future calendar_years.

Examples

# for a constant inflation of 2%
historic_indices <- data.frame(Calendar_year = 2015:2023,
                               Index_gross = 0.03)
print(historic_indices)
#>   Calendar_year Index_gross
#> 1          2015        0.03
#> 2          2016        0.03
#> 3          2017        0.03
#> 4          2018        0.03
#> 5          2019        0.03
#> 6          2020        0.03
#> 7          2021        0.03
#> 8          2022        0.03
#> 9          2023        0.03
expand_historic_indices(historic_indices = historic_indices,
                        first_orig_year = 2015,
                        last_orig_year = 2023,
                        index_gross_future = 0.03)
#>     Calendar_year Index_gross Index_re
#> 1            2015        0.03        0
#> 2            2016        0.03        0
#> 3            2017        0.03        0
#> 4            2018        0.03        0
#> 5            2019        0.03        0
#> 6            2020        0.03        0
#> 7            2021        0.03        0
#> 8            2022        0.03        0
#> 9            2023        0.03        0
#> 10           2024        0.03        0
#> 11           2025        0.03        0
#> 12           2026        0.03        0
#> 13           2027        0.03        0
#> 14           2028        0.03        0
#> 15           2029        0.03        0
#> 16           2030        0.03        0
#> 17           2031        0.03        0
#> 18           2032        0.03        0
#> 19           2033        0.03        0
#> 20           2034        0.03        0
#> 21           2035        0.03        0
#> 22           2036        0.03        0
#> 23           2037        0.03        0
#> 24           2038        0.03        0
#> 25           2039        0.03        0
#> 26           2040        0.03        0
#> 27           2041        0.03        0
#> 28           2042        0.03        0
#> 29           2043        0.03        0
#> 30           2044        0.03        0
#> 31           2045        0.03        0
#> 32           2046        0.03        0
#> 33           2047        0.03        0
#> 34           2048        0.03        0
#> 35           2049        0.03        0
#> 36           2050        0.03        0
#> 37           2051        0.03        0
#> 38           2052        0.03        0
#> 39           2053        0.03        0
#> 40           2054        0.03        0
#> 41           2055        0.03        0
#> 42           2056        0.03        0
#> 43           2057        0.03        0
#> 44           2058        0.03        0
#> 45           2059        0.03        0
#> 46           2060        0.03        0
#> 47           2061        0.03        0
#> 48           2062        0.03        0
#> 49           2063        0.03        0
#> 50           2064        0.03        0
#> 51           2065        0.03        0
#> 52           2066        0.03        0
#> 53           2067        0.03        0
#> 54           2068        0.03        0
#> 55           2069        0.03        0
#> 56           2070        0.03        0
#> 57           2071        0.03        0
#> 58           2072        0.03        0
#> 59           2073        0.03        0
#> 60           2074        0.03        0
#> 61           2075        0.03        0
#> 62           2076        0.03        0
#> 63           2077        0.03        0
#> 64           2078        0.03        0
#> 65           2079        0.03        0
#> 66           2080        0.03        0
#> 67           2081        0.03        0
#> 68           2082        0.03        0
#> 69           2083        0.03        0
#> 70           2084        0.03        0
#> 71           2085        0.03        0
#> 72           2086        0.03        0
#> 73           2087        0.03        0
#> 74           2088        0.03        0
#> 75           2089        0.03        0
#> 76           2090        0.03        0
#> 77           2091        0.03        0
#> 78           2092        0.03        0
#> 79           2093        0.03        0
#> 80           2094        0.03        0
#> 81           2095        0.03        0
#> 82           2096        0.03        0
#> 83           2097        0.03        0
#> 84           2098        0.03        0
#> 85           2099        0.03        0
#> 86           2100        0.03        0
#> 87           2101        0.03        0
#> 88           2102        0.03        0
#> 89           2103        0.03        0
#> 90           2104        0.03        0
#> 91           2105        0.03        0
#> 92           2106        0.03        0
#> 93           2107        0.03        0
#> 94           2108        0.03        0
#> 95           2109        0.03        0
#> 96           2110        0.03        0
#> 97           2111        0.03        0
#> 98           2112        0.03        0
#> 99           2113        0.03        0
#> 100          2114        0.03        0
#> 101          2115        0.03        0
#> 102          2116        0.03        0
#> 103          2117        0.03        0
#> 104          2118        0.03        0
#> 105          2119        0.03        0
#> 106          2120        0.03        0
#> 107          2121        0.03        0
#> 108          2122        0.03        0
#> 109          2123        0.03        0
#> 110          2124        0.03        0
#> 111          2125        0.03        0
#> 112          2126        0.03        0
#> 113          2127        0.03        0
#> 114          2128        0.03        0
#> 115          2129        0.03        0
#> 116          2130        0.03        0
#> 117          2131        0.03        0
#> 118          2132        0.03        0
#> 119          2133        0.03        0
#> 120          2134        0.03        0
#> 121          2135        0.03        0
#> 122          2136        0.03        0
#> 123          2137        0.03        0
#> 124          2138        0.03        0
#> 125          2139        0.03        0
#> 126          2140        0.03        0
#> 127          2141        0.03        0
#> 128          2142        0.03        0
#> 129          2143        0.03        0
#> 130          2144        0.03        0
#> 131          2145        0.03        0
#> 132          2146        0.03        0
#> 133          2147        0.03        0
#> 134          2148        0.03        0
#> 135          2149        0.03        0
#> 136          2150        0.03        0
#> 137          2151        0.03        0
#> 138          2152        0.03        0
#> 139          2153        0.03        0
#> 140          2154        0.03        0
#> 141          2155        0.03        0
#> 142          2156        0.03        0
#> 143          2157        0.03        0
#> 144          2158        0.03        0
#> 145          2159        0.03        0
#> 146          2160        0.03        0
#> 147          2161        0.03        0
#> 148          2162        0.03        0
#> 149          2163        0.03        0
#> 150          2164        0.03        0
#> 151          2165        0.03        0
#> 152          2166        0.03        0
#> 153          2167        0.03        0
#> 154          2168        0.03        0
#> 155          2169        0.03        0
#> 156          2170        0.03        0
#> 157          2171        0.03        0
#> 158          2172        0.03        0
#> 159          2173        0.03        0
#> 160          2174        0.03        0
#> 161          2175        0.03        0
#> 162          2176        0.03        0
#> 163          2177        0.03        0
#> 164          2178        0.03        0
#> 165          2179        0.03        0
#> 166          2180        0.03        0
#> 167          2181        0.03        0
#> 168          2182        0.03        0
#> 169          2183        0.03        0
#> 170          2184        0.03        0
#> 171          2185        0.03        0
#> 172          2186        0.03        0
#> 173          2187        0.03        0
#> 174          2188        0.03        0
#> 175          2189        0.03        0
#> 176          2190        0.03        0
#> 177          2191        0.03        0
#> 178          2192        0.03        0
#> 179          2193        0.03        0
#> 180          2194        0.03        0
#> 181          2195        0.03        0
#> 182          2196        0.03        0
#> 183          2197        0.03        0
#> 184          2198        0.03        0
#> 185          2199        0.03        0
#> 186          2200        0.03        0
#> 187          2201        0.03        0
#> 188          2202        0.03        0
#> 189          2203        0.03        0
#> 190          2204        0.03        0
#> 191          2205        0.03        0
#> 192          2206        0.03        0
#> 193          2207        0.03        0
#> 194          2208        0.03        0
#> 195          2209        0.03        0
#> 196          2210        0.03        0
#> 197          2211        0.03        0
#> 198          2212        0.03        0
#> 199          2213        0.03        0
#> 200          2214        0.03        0
#> 201          2215        0.03        0
#> 202          2216        0.03        0
#> 203          2217        0.03        0
#> 204          2218        0.03        0
#> 205          2219        0.03        0
#> 206          2220        0.03        0
#> 207          2221        0.03        0
#> 208          2222        0.03        0
#> 209          2223        0.03        0
#> 210          2224        0.03        0
#> 211          2225        0.03        0
#> 212          2226        0.03        0
#> 213          2227        0.03        0
#> 214          2228        0.03        0
#> 215          2229        0.03        0
#> 216          2230        0.03        0
#> 217          2231        0.03        0
#> 218          2232        0.03        0
#> 219          2233        0.03        0
#> 220          2234        0.03        0
#> 221          2235        0.03        0
#> 222          2236        0.03        0
#> 223          2237        0.03        0
#> 224          2238        0.03        0
#> 225          2239        0.03        0
#> 226          2240        0.03        0
#> 227          2241        0.03        0
#> 228          2242        0.03        0
#> 229          2243        0.03        0
#> 230          2244        0.03        0
#> 231          2245        0.03        0
#> 232          2246        0.03        0
#> 233          2247        0.03        0
#> 234          2248        0.03        0
#> 235          2249        0.03        0
#> 236          2250        0.03        0
#> 237          2251        0.03        0
#> 238          2252        0.03        0
#> 239          2253        0.03        0
#> 240          2254        0.03        0
#> 241          2255        0.03        0
#> 242          2256        0.03        0
#> 243          2257        0.03        0
#> 244          2258        0.03        0
#> 245          2259        0.03        0
#> 246          2260        0.03        0
#> 247          2261        0.03        0
#> 248          2262        0.03        0
#> 249          2263        0.03        0
#> 250          2264        0.03        0
#> 251          2265        0.03        0
#> 252          2266        0.03        0
#> 253          2267        0.03        0
#> 254          2268        0.03        0
#> 255          2269        0.03        0
#> 256          2270        0.03        0
#> 257          2271        0.03        0
#> 258          2272        0.03        0
#> 259          2273        0.03        0