Generate claims list

The prepared data will be reduced to claims with origin year between first_orig_year and last_orig_year and to the calendar year last_orig_year. Only necessary columns will be returned.

Usage

generate_claims_list(extended_claims_data, first_orig_year, last_orig_year)

Arguments

extended_claims_data

prepared dataframe, see details of prepare_data().

first_orig_year

Integer value for the first origin year with complete history.

last_orig_year

Integer value for the last origin year which will also be treated as the last calendar year, usually the calendar year just ended.

Value

dataframe with current large claim status

Examples

# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)
claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                    first_orig_year = 1989,
                                    last_orig_year = 2023)
print(claims_list)
#>    Claim_id Origin_year Exit_reserve_class Dev_year_since_large
#> 1  Claim#10        1989                  1                   10
#> 2  Claim#13        1990                  1                   34
#> 3  Claim#16        1993                  0                   31
#> 4  Claim#22        1996                  1                   25
#> 5  Claim#32        2000                  1                   20
#> 6  Claim#37        2003                  0                   21
#> 7  Claim#41        2005                  0                   18
#> 8  Claim#43        2006                  2                   17
#> 9  Claim#44        2006                  3                   17
#> 10 Claim#45        2006                  2                   18
#> 11 Claim#49        2010                  0                   13
#> 12 Claim#51        2010                  1                    7
#> 13 Claim#52        2012                  2                   12
#> 14 Claim#60        2015                  1                    6
#> 15 Claim#68        2019                  2                    2
#>    Dev_year_of_growing_large Large_since Cl_reserve Ind_cl_reserve
#> 1                         26        2014   14396.99       13588.93
#> 2                          1        1990  114612.16      108179.34
#> 3                          1        1993       0.00           0.00
#> 4                          4        1999   89286.40       84275.04
#> 5                          5        2004  145280.12      137126.01
#> 6                          1        2003       0.00           0.00
#> 7                          2        2006       0.00           0.00
#> 8                          2        2007  240134.33      226656.36
#> 9                          2        2007  514406.97      485534.94
#> 10                         1        2006  268343.58      253282.31
#> 11                         2        2011       0.00           0.00
#> 12                         8        2017  160590.46      151577.03
#> 13                         1        2012  214846.87      202788.20
#> 14                         4        2018   27561.04       26014.13
#> 15                         4        2022  289074.16      272849.35