Generate ibnr pools

For sampling purposes this pool shall contain a list with one vector per development year. The vector contains historic ibnr claims.

Usage

generate_ibnr_pools(
  large_claims_list,
  first_orig_year,
  last_orig_year,
  years_for_ibnr_pools
)

Arguments

large_claims_list

Dataframe with one row per known large claim generated by generate_claims_list().

first_orig_year

First origin year with complete history.

last_orig_year

Last origin year.

years_for_ibnr_pools

Vector containing the calendar years that shall be used to build the pools. Thus ibnr claims that have become large in other calendar years don't affect the model.

Value

List of one vector per development year with each vector containing the row numbers of the claims in large_claims_list that have become large in that development year and in a calendar year in years_for_ibnr_pools.

Examples

# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)
generate_ibnr_pools(large_claims_list_xmpl, 1989, 2023, 2014:2023)
#> [[1]]
#>  [1] 618 619 625 630 641 646 647 648 651 653 658 660 661 664 670 675 683 690 692
#> [20] 697 702 706 708 709 712 718 722 727 739 740 742 744 753 755 756 763 764 765
#> [39] 778 779 781 784
#> 
#> [[2]]
#>  [1] 592 604 634 668 681 717 719 730 734 743 750 758 769 772 773
#> 
#> [[3]]
#> [1] 566 594 613 638 699 700 752 754 768
#> 
#> [[4]]
#> [1] 572 612 632 639 687
#> 
#> [[5]]
#> [1] 521 553 561 584 611 616 620 635
#> 
#> [[6]]
#> [1] 527 552 576 621
#> 
#> [[7]]
#> [1] 472 493 680
#> 
#> [[8]]
#> [1] 530 591
#> 
#> [[9]]
#> [1] 458 544
#> 
#> [[10]]
#> integer(0)
#> 
#> [[11]]
#> [1] 391 491
#> 
#> [[12]]
#> integer(0)
#> 
#> [[13]]
#> integer(0)
#> 
#> [[14]]
#> [1] 480
#> 
#> [[15]]
#> [1] 373
#> 
#> [[16]]
#> integer(0)
#> 
#> [[17]]
#> integer(0)
#> 
#> [[18]]
#> integer(0)
#> 
#> [[19]]
#> integer(0)
#> 
#> [[20]]
#> [1] 135
#> 
#> [[21]]
#> [1] 231 245
#> 
#> [[22]]
#> integer(0)
#> 
#> [[23]]
#> integer(0)
#> 
#> [[24]]
#> integer(0)
#> 
#> [[25]]
#> integer(0)
#> 
#> [[26]]
#> [1] 41
#> 
#> [[27]]
#> integer(0)
#> 
#> [[28]]
#> integer(0)
#> 
#> [[29]]
#> integer(0)
#> 
#> [[30]]
#> integer(0)
#> 
#> [[31]]
#> integer(0)
#> 
#> [[32]]
#> integer(0)
#> 
#> [[33]]
#> integer(0)
#> 
#> [[34]]
#> integer(0)
#>