Generating additive model to estimate future ibnr claim numbers

This function calculates the number of ibnr claims per development year and divides it by the exposure these ibnr claims have raised from.

Usage

get_additive_ibnr_model(
  large_claims_list,
  first_orig_year,
  last_orig_year,
  exposure,
  years_for_ibnr_pools
)

Arguments

large_claims_list

Dataframe with one row per known large claim as a result of generate_claims_list()

first_orig_year

Desired first origin year.

last_orig_year

Desired last origin year.

exposure

Dataframe that must contain one row for each origin year between first_orig_year and last_orig_year and the columns Origin_year and Exposure.

years_for_ibnr_pools

Vector containing the calendar years that shall be used to build the pools. Thus ibnr claims that have became large in other calendar years don't affect the model.

Value

Dataframe with columns Dev_year, Ibnr_numbers, Sum_of_exposures and Factor.

Examples

# prepare data
extended_claims_data_xmpl <- prepare_data(claims_data = claims_data_xmpl,
                                          indices = indices_xmpl,
                                          threshold = 4e5,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023,
                                          expected_year_of_growing_large = 3,
                                          reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                          pool_of_annuities = pool_of_annuities_xmpl)

# generate claims list
large_claims_list_xmpl <- generate_claims_list(extended_claims_data = extended_claims_data_xmpl,
                                               first_orig_year = 1989,
                                               last_orig_year = 2023)

# generate additive ibnr model using the last ten calendar years
get_additive_ibnr_model(large_claims_list = large_claims_list_xmpl,
                        first_orig_year = 1989,
                        last_orig_year = 2023,
                        exposure = exposure_xmpl,
                        years_for_ibnr_pools = 2014:2023)
#>    Dev_year Ibnr_numbers Sum_of_exposures       Factor
#> 1         1           96       3251278000 2.952685e-08
#> 2         2           42       3186814000 1.317931e-08
#> 3         3           15       3135936000 4.783261e-09
#> 4         4            9       3091159000 2.911529e-09
#> 5         5            5       3052563000 1.637968e-09
#> 6         6            8       3018087000 2.650686e-09
#> 7         7            4       2988493000 1.338467e-09
#> 8         8            3       2960377000 1.013384e-09
#> 9         9            2       2929769000 6.826477e-10
#> 10       10            2       2900091000 6.896335e-10
#> 11       11            0       2871908000 0.000000e+00
#> 12       12            2       2843332000 7.034001e-10
#> 13       13            0       2814408000 0.000000e+00
#> 14       14            0       2784593000 0.000000e+00
#> 15       15            1       2753571000 3.631648e-10
#> 16       16            1       2722377000 3.673261e-10
#> 17       17            0       2688906000 0.000000e+00
#> 18       18            0       2651047000 0.000000e+00
#> 19       19            0       2617325000 0.000000e+00
#> 20       20            0       2585507000 0.000000e+00
#> 21       21            1       2551507000 3.919252e-10
#> 22       22            2       2516211000 7.948459e-10
#> 23       23            0       2476928000 0.000000e+00
#> 24       24            0       2433118000 0.000000e+00
#> 25       25            0       2371300000 0.000000e+00
#> 26       26            0       2306931000 0.000000e+00
#> 27       27            1       2054156000 4.868179e-10
#> 28       28            0       1807777000 0.000000e+00
#> 29       29            0       1564193000 0.000000e+00
#> 30       30            0       1321361000 0.000000e+00
#> 31       31            0       1082358000 0.000000e+00
#> 32       32            0        846544000 0.000000e+00
#> 33       33            0        616467000 0.000000e+00
#> 34       34            0        394126000 0.000000e+00