Helper function to get start, sample_size and range of a pool in the sample_matrix

The function pools2matrix() converts the pools into a matrix for faster simulation. This function helps to find the start and end of a pool in the matrix. For description check details of pools2matrix().

Usage

get_matrix_info(
  dev_year,
  reserve_class,
  start_of_tail,
  number_of_reserve_classes,
  lengths_vector,
  output
)

Arguments

dev_year

Development year of desired pool.

reserve_class

Reserve class of desired pool.

start_of_tail

Start of tail.

number_of_reserve_classes

Number of reserve classes.

lengths_vector

Vector containing the lengths of each pool (first entry of list output of pools2matrix()).

output

One of sample_size, start or range.

Value

Integer stating the row number if output equals sample_size or start, integer vector of row numbers in case of range.

Examples

if (FALSE) { # \dontrun{
# this example uses data provided with this package
new_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                         indices = indices_xmpl,
                         threshold = 400000,
                         first_orig_year = 1989,
                         last_orig_year = 2023,
                         expected_year_of_growing_large = 3,
                         reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                         pool_of_annuities = minimal_pool_of_annuities_xmpl)
pools <- generate_pools(extended_claims_data = new_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = minimal_pool_of_annuities_xmpl)
pools_as_matrix <- pools2matrix(pools)
get_matrix_info(10, 2, 17, 6, pools_as_matrix[[1]], "start")
get_matrix_info(10, 2, 17, 6, pools_as_matrix[[1]], "sample_size")
get_matrix_info(10, 2, 17, 6, pools_as_matrix[[1]], "range")} # }