Example of a indices dataframe

Dataframe contains one row for each historic calendar year and 250 future calendar years.
This dataframe emerges from the dataframe historic_indices_xmpl and an expected future constant gross index of 3% and an expected future constant reinsurance index of 2%.

Usage

indices_xmpl

Format

indices_xmpl

A data frame with 286 rows and 4 columns:

Calendar_year

Calendar year

Index_gross

is the claim payment development from one year to another, e.g. 0.02 for 2% increase

Index_re

is the contractually fixed claim payment development that is to be used in special index clauses that are a common part of longtail xl resinsurance programs.

Transition_factor

The transition_factor for year j indexes a payment of year j to the niveau of payments in the fixed index_year.

Source

Dataframe indices_xmpl can be generated by:

library(dplyr)
indices_xmpl <- expand_historic_indices(historic_indices_xmpl,
                                        first_orig_year = 1989,
                                        last_orig_year = 2023,
                                        index_gross_future = 0.03,
                                        index_re_future = 0.02) %>%
                add_transition_factor(index_year = 2022)