merge paid and incurred results

merging the results of paid and incurred triangle methods with given weights

Usage

merge_results(
  paid_result,
  incurred_result,
  weight_paid = 1,
  weight_incurred = 1
)

Arguments

paid_result

ultimate vector or filled up cumulative paid claims triangle. numeric vector or numeric matrix

incurred_result

ultimate vector or filled up cumulative incurred claims triangle. numeric vector or numeric matrix

weight_paid

Default: 1. numeric vector of length 1 or same as number of rows of paid_result

weight_incurred

Default: 1. numeric vector of length 1 or same as number of rows of incurred_result

Value

numeric vector

Examples

pai <- c(100, 200, 300)
filled_paid <- matrix(c(70, 150, 220, 90, 180, 270, 100, 200, 300), 3)
inc <- c(110, 190, 290)
merge_results(pai, inc, 1, 1)
#> [1] 105 195 295
merge_results(pai, inc, c(2,1,0), 1)
#> [1] 103.3333 195.0000 290.0000
merge_results(filled_paid, inc, c(2,1,0), 1)
#> [1] 103.3333 195.0000 290.0000