Minimal example of a claims data dataframe

Dataframe contains one row for each calendar year per claim where at least one of the columns Cl_payment_cal, Cl_reserve, An_payment_cal, An_reserve is not 0.
Data belongs to the set of minimal_claims_data_xmpl, minimal_active_annuities_xmpl and minimal_pool_of_annuities_xmpl.

Usage

minimal_claims_data_xmpl

Format

minimal_claims_data_xmpl

A data frame with 995 rows and 7 columns:

Claim_id

Claim ID

Origin_year

Origin year of the claim

Calendar_year

Calendar year

Cl_payment_cal

Claim payments without annuities

Cl_reserve

Claim payments without annuities

An_payment_cal

Annuities' payments

An_reserve

Annuities' reserve

Source

A subset derived by reducing and anonymising an original dataset of an insurer.