Minimal example of a pool of annuities dataframe

Dataframe contains one row for each annuity that has ever been agreed on regardless of whether the annuity is still active or not
Data belongs to the set of minimal_claims_data_xmpl, minimal_active_annuities_xmpl and minimal_pool_of_annuities_xmpl.

Usage

minimal_pool_of_annuities_xmpl

Format

minimal_pool_of_annuities_xmpl

A data frame with 3 rows and 11 columns:

Claim_id

Claim ID

Annuity_id

Annuity ID for the case of multiple annuities in one claim

Origin_year

Origin year of the claim

Calendar_year

Calendar year

Entering_year

Year in which insurer and recipient have agreed on the annuity

Annuity_start

Year in which payment starts, may be a past or a future year

Annuity_end

Year in which payment ends, may be a past or a future year

Birth_year

Birth year of recipient for assigning survival probability from mortality tables

Gender

Gender of recipient for assigning mortality table

Annual_payment

Agreed annual payment

Dynamic

If a dynamic increase of the annual payment is part of the agreement, it can be specified here, e.g. 0.02 for 2% annual payment increasement

Source

A subset derived by reducing and anonymising an original dataset of an insurer.