Converts pools into a dataframe

This function converts the first two of the three list entries of the object pools into an easier to read dataframe.

Usage

pools2dataframe(pools)

Arguments

pools

List as output of generate_pools().

Value

Dataframe with the entries of pools plus the columns Dev_year_since_large and Entry_reserve_class.

Examples

# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = minimal_pool_of_annuities_xmpl)

pools_as_df <- pools2dataframe(pools)