Reduce claims data to possible large claims

The original dataframe is reduced to possible large claims to save computing time in the next steps.

This is not yet a precise separation as the indexed columns are missing at this point.

This function resembles only step 1 of the data preparation process, see description in details of prepare_data()

Usage

reduce_data(claims_data, indices, threshold)

Arguments

claims_data

dataframe, see details of prepare_data().

indices

dataframe, see details of prepare_data().

threshold

numeric value to separate small from large claims, see details of prepare_data().

Value

subset of dataframe claims_data where preselected small claims have been eliminated

Examples

# this example uses data provided with this package
reduce_data(minimal_claims_data_xmpl, indices_xmpl, threshold = 400000)
#>     Claim_id Origin_year Calendar_year Cl_payment_cal   Cl_reserve
#> 1    Claim#1        1976          1988    74818.22754   25347.5130
#> 2    Claim#1        1976          1989    10994.54132   25347.5130
#> 3    Claim#1        1976          1990     8253.98220   25347.5130
#> 4    Claim#1        1976          1991     1342.11132   25347.5130
#> 5    Claim#1        1976          1992     6148.26862   25347.5130
#> 6    Claim#1        1976          1993      574.60542   25347.5130
#> 7    Claim#1        1976          1994    11657.40136   25347.5130
#> 8    Claim#1        1976          1995      829.09283   25347.5130
#> 9    Claim#1        1976          1996      488.08952   24926.1285
#> 10   Claim#1        1976          1997     5302.95388   19271.6452
#> 11   Claim#1        1976          1998      815.22182  100456.0656
#> 12   Claim#1        1976          1999      339.30482   99985.1065
#> 13   Claim#1        1976          2000    77411.34212   31177.4905
#> 14   Claim#1        1976          2001      384.38303   30756.1061
#> 15   Claim#1        1976          2002      368.85309   30344.6366
#> 16   Claim#1        1976          2003      487.69172   29870.7030
#> 17   Claim#1        1976          2004      452.97321   29440.3952
#> 18   Claim#1        1976          2005      501.02662   28972.4106
#> 19   Claim#1        1976          2006      464.57064   28414.2001
#> 20   Claim#1        1976          2007      452.80804   27946.2155
#> 21   Claim#1        1976          2008     7356.75644   19298.4155
#> 22   Claim#1        1976          2009        0.00000   19298.4155
#> 23   Claim#1        1976          2010        0.00000   19298.4155
#> 24   Claim#1        1976          2011        0.00000   19298.4155
#> 25   Claim#1        1976          2012        0.00000   19298.4155
#> 26   Claim#1        1976          2013        0.00000   19298.4155
#> 27   Claim#1        1976          2014        0.00000   19298.4155
#> 28   Claim#1        1976          2015        0.00000   19298.4155
#> 29   Claim#1        1976          2016        0.00000   19298.4155
#> 30   Claim#1        1976          2017        0.00000   19298.4155
#> 31   Claim#1        1976          2018        0.00000   19298.4155
#> 32   Claim#1        1976          2019        0.00000   49574.6391
#> 33   Claim#1        1976          2020        0.00000   49574.6391
#> 34   Claim#1        1976          2021        0.00000   49574.6391
#> 35   Claim#1        1976          2022        0.00000   49574.6391
#> 36   Claim#1        1976          2023        0.00000   49574.6391
#> 37   Claim#4        1982          1988   459071.74452  250327.6488
#> 38   Claim#4        1982          1989    47281.39933  250327.6488
#> 39   Claim#4        1982          1990      880.35899  250327.6488
#> 40   Claim#4        1982          1991    46619.07120  250327.6488
#> 41   Claim#4        1982          1992     7391.29956  250327.6488
#> 42   Claim#4        1982          1993     8085.81188  250327.6488
#> 43   Claim#4        1982          1994    20504.50143  166885.8246
#> 44   Claim#4        1982          1995    10591.89088  152493.7881
#> 45   Claim#4        1982          1996        0.00000  152493.7881
#> 46   Claim#4        1982          1997        0.00000  152493.7881
#> 47   Claim#4        1982          1998   113462.79266       0.0000
#> 48   Claim#5        1984          1988   154112.66640       0.0000
#> 49   Claim#5        1984          1989    63171.04934       0.0000
#> 50   Claim#6        1985          1988    14930.75139    4795.9425
#> 51   Claim#6        1985          1989     1085.17193    4795.9425
#> 52   Claim#6        1985          1990     5645.97159   23979.7127
#> 53   Claim#6        1985          1991        0.00000   23979.7127
#> 54   Claim#6        1985          1992        0.00000   57550.3725
#> 55   Claim#6        1985          1993     4328.50122   47958.4874
#> 56   Claim#6        1985          1994        0.00000   47958.4874
#> 57   Claim#6        1985          1995        0.00000   47958.4874
#> 58   Claim#6        1985          1996        0.00000   47958.4874
#> 59   Claim#6        1985          1997        0.00000   47958.4874
#> 60   Claim#6        1985          1998        0.00000   47958.4874
#> 61   Claim#6        1985          1999        0.00000  239793.3751
#> 62   Claim#6        1985          2000        0.00000  239793.3751
#> 63   Claim#6        1985          2001        0.00000  239793.3751
#> 64   Claim#6        1985          2002        0.00000  239793.3751
#> 65   Claim#6        1985          2003        0.00000  239793.3751
#> 66   Claim#6        1985          2004        0.00000  239793.3751
#> 67   Claim#6        1985          2005        0.00000  239793.3751
#> 68   Claim#6        1985          2006        0.00000  239793.3751
#> 69   Claim#6        1985          2007        0.00000  239793.3751
#> 70   Claim#6        1985          2008        0.00000  239793.3751
#> 71   Claim#6        1985          2009        0.00000  239793.3751
#> 72   Claim#6        1985          2010        0.00000  239793.3751
#> 73   Claim#6        1985          2011        0.00000  239793.3751
#> 74   Claim#6        1985          2012        0.00000  239793.3751
#> 75   Claim#6        1985          2013        0.00000  239793.3751
#> 76   Claim#6        1985          2014        0.00000  239793.3751
#> 77   Claim#6        1985          2015        0.00000  239793.3751
#> 78   Claim#6        1985          2016        0.00000  239793.3751
#> 79   Claim#6        1985          2017    73954.91808   33077.2771
#> 80   Claim#6        1985          2018        0.00000   33077.2771
#> 81   Claim#6        1985          2019        0.00000   33077.2771
#> 82  Claim#10        1989          1989    27723.33481   41627.7654
#> 83  Claim#10        1989          1990     3562.17950   83255.5308
#> 84  Claim#10        1989          1991      476.60328   74930.6290
#> 85  Claim#10        1989          1992     1582.56257   74930.6290
#> 86  Claim#10        1989          1993      834.38355   74930.6290
#> 87  Claim#10        1989          1994     3396.17587   74930.6290
#> 88  Claim#10        1989          1995    27510.22188   51440.9723
#> 89  Claim#10        1989          1996     1597.74691   50192.8477
#> 90  Claim#10        1989          1997    27173.21487   83015.3502
#> 91  Claim#10        1989          1998        0.00000   83015.3502
#> 92  Claim#10        1989          1999        0.00000   83015.3502
#> 93  Claim#10        1989          2000    50817.87082   43368.4637
#> 94  Claim#10        1989          2001        0.00000   43368.4637
#> 95  Claim#10        1989          2002        0.00000   43368.4637
#> 96  Claim#10        1989          2003        0.00000   43368.4637
#> 97  Claim#10        1989          2004        0.00000   43368.4637
#> 98  Claim#10        1989          2005        0.00000   43368.4637
#> 99  Claim#10        1989          2006        0.00000   43368.4637
#> 100 Claim#10        1989          2007        0.00000   43368.4637
#> 101 Claim#10        1989          2008        0.00000   43368.4637
#> 102 Claim#10        1989          2009        0.00000   43368.4637
#> 103 Claim#10        1989          2010        0.00000   43368.4637
#> 104 Claim#10        1989          2011        0.00000   43368.4637
#> 105 Claim#10        1989          2012    18770.07288   23576.7731
#> 106 Claim#10        1989          2013     4745.00001  118320.2612
#> 107 Claim#10        1989          2014   167320.14251   16283.4264
#> 108 Claim#10        1989          2015        0.00000   16283.4264
#> 109 Claim#10        1989          2016        0.00000   16283.4264
#> 110 Claim#10        1989          2017        0.00000   16283.4264
#> 111 Claim#10        1989          2018        0.00000   16283.4264
#> 112 Claim#10        1989          2019        0.00000   16283.4264
#> 113 Claim#10        1989          2020        0.00000   16283.4264
#> 114 Claim#10        1989          2021        0.00000   16283.4264
#> 115 Claim#10        1989          2022     1840.90083   14396.9914
#> 116 Claim#10        1989          2023        0.00000   14396.9914
#> 117 Claim#13        1990          1990    15662.59994  239728.0056
#> 118 Claim#13        1990          1991    74056.30973  346274.5962
#> 119 Claim#13        1990          1992    87205.07678  239728.0056
#> 120 Claim#13        1990          1993    12714.46615  239728.0056
#> 121 Claim#13        1990          1994     8034.93740  239728.0056
#> 122 Claim#13        1990          1995      899.71696  239728.0056
#> 123 Claim#13        1990          1996     5913.43845  233588.9615
#> 124 Claim#13        1990          1997     6699.12009  224195.9739
#> 125 Claim#13        1990          1998      747.18531  202437.3761
#> 126 Claim#13        1990          1999    10021.61237  188059.8015
#> 127 Claim#13        1990          2000     9158.19674  179552.4533
#> 128 Claim#13        1990          2001      423.65566  179106.5078
#> 129 Claim#13        1990          2002     8245.81400  169367.5999
#> 130 Claim#13        1990          2003      344.63245  302028.0505
#> 131 Claim#13        1990          2004     9141.82922  249750.3181
#> 132 Claim#13        1990          2005     7493.26499  239743.6345
#> 133 Claim#13        1990          2006     9835.20073  229699.4415
#> 134 Claim#13        1990          2007    54163.43250  333417.1956
#> 135 Claim#13        1990          2008    15100.60063  316302.4742
#> 136 Claim#13        1990          2009    11219.22501  303531.5537
#> 137 Claim#13        1990          2010     1952.67950  300248.4362
#> 138 Claim#13        1990          2011     3512.94297  296537.0861
#> 139 Claim#13        1990          2012    29562.17384  253924.2846
#> 140 Claim#13        1990          2013    12046.37383  241712.8798
#> 141 Claim#13        1990          2014    29905.53968  211204.1645
#> 142 Claim#13        1990          2015        0.00000  211204.1645
#> 143 Claim#13        1990          2016        0.00000  211204.1645
#> 144 Claim#13        1990          2017        0.00000  211204.1645
#> 145 Claim#13        1990          2018        0.00000  125031.4484
#> 146 Claim#13        1990          2019        0.00000  125031.4484
#> 147 Claim#13        1990          2020        0.00000  125031.4484
#> 148 Claim#13        1990          2021        0.00000  125031.4484
#> 149 Claim#13        1990          2022        0.00000  125031.4484
#> 150 Claim#13        1990          2023        0.00000  114612.1610
#> 151 Claim#16        1993          1993    47901.86000  246654.9603
#> 152 Claim#16        1993          1994   222066.49472  318595.9569
#> 153 Claim#16        1993          1995    24208.01902  303407.1088
#> 154 Claim#16        1993          1996    32412.14068  273461.1745
#> 155 Claim#16        1993          1997   101757.26256  164090.3732
#> 156 Claim#16        1993          1998        0.00000  164090.3732
#> 157 Claim#16        1993          1999        0.00000  164090.3732
#> 158 Claim#16        1993          2000        0.00000  164090.3732
#> 159 Claim#16        1993          2001        0.00000  164090.3732
#> 160 Claim#16        1993          2002        0.00000  164090.3732
#> 161 Claim#16        1993          2003        0.00000  164090.3732
#> 162 Claim#16        1993          2004     5629.32132       0.0000
#> 163 Claim#17        1994          1994        0.00000   19185.7175
#> 164 Claim#17        1994          1995    19087.55857   38370.4970
#> 165 Claim#17        1994          1996    10688.47419  190036.3521
#> 166 Claim#17        1994          1997    24432.83314  164209.9298
#> 167 Claim#17        1994          1998        0.00000   23982.1469
#> 168 Claim#17        1994          1999        0.00000   23982.1469
#> 169 Claim#17        1994          2000       89.59402   23888.3384
#> 170 Claim#17        1994          2001        0.00000   23888.3384
#> 171 Claim#17        1994          2002       79.03382   23818.9201
#> 172 Claim#17        1994          2003        0.00000   23818.9201
#> 173 Claim#17        1994          2004      214.51231   23587.2131
#> 174 Claim#17        1994          2005        0.00000   23587.2131
#> 175 Claim#17        1994          2006      448.92760   23189.4650
#> 176 Claim#17        1994          2007      209.90955   22996.2194
#> 177 Claim#17        1994          2008        0.00000   22996.2194
#> 178 Claim#17        1994          2009      321.36999   22705.4130
#> 179 Claim#17        1994          2010        0.00000   22705.4130
#> 180 Claim#17        1994          2011       82.97936   22605.0379
#> 181 Claim#17        1994          2012     2630.62736   19790.7825
#> 182 Claim#17        1994          2013        0.00000   19790.7825
#> 183 Claim#17        1994          2014        0.00000   19790.7825
#> 184 Claim#17        1994          2015        0.00000   19790.7825
#> 185 Claim#17        1994          2016        0.00000   19790.7825
#> 186 Claim#17        1994          2017        0.00000   19790.7825
#> 187 Claim#17        1994          2018        0.00000   19790.7825
#> 188 Claim#17        1994          2019        0.00000   19790.7825
#> 189 Claim#17        1994          2020        0.00000   19790.7825
#> 190 Claim#17        1994          2021        0.00000   19790.7825
#> 191 Claim#21        1996          1996      311.97025  138323.9022
#> 192 Claim#21        1996          1997    22932.77509  193923.7846
#> 193 Claim#21        1996          1998    19332.74087  108267.1332
#> 194 Claim#21        1996          1999    16534.81474   90980.8388
#> 195 Claim#21        1996          2000        0.00000   90980.8388
#> 196 Claim#21        1996          2001        0.00000   90980.8388
#> 197 Claim#21        1996          2002        0.00000   90980.8388
#> 198 Claim#21        1996          2003        0.00000   90980.8388
#> 199 Claim#21        1996          2004        0.00000   90980.8388
#> 200 Claim#21        1996          2005        0.00000   90980.8388
#> 201 Claim#21        1996          2006        0.00000   90980.8388
#> 202 Claim#22        1996          1996     5718.35528   27760.7665
#> 203 Claim#22        1996          1997    47157.58670  188796.9134
#> 204 Claim#22        1996          1998    43129.15144  151868.0566
#> 205 Claim#22        1996          1999    27089.34746  544733.9169
#> 206 Claim#22        1996          2000    74668.48781  472164.3626
#> 207 Claim#22        1996          2001     6433.72349  466744.6779
#> 208 Claim#22        1996          2002     6092.30408  461739.7691
#> 209 Claim#22        1996          2003     7356.86764  456346.8703
#> 210 Claim#22        1996          2004     7472.91845   64935.5667
#> 211 Claim#22        1996          2005      333.19867   64649.8502
#> 212 Claim#22        1996          2006    10562.94010   56220.4019
#> 213 Claim#22        1996          2007        0.00000   56220.4019
#> 214 Claim#22        1996          2008      126.34484   52760.1479
#> 215 Claim#22        1996          2009     1955.60207   51081.5635
#> 216 Claim#22        1996          2010     1773.97313   49722.7868
#> 217 Claim#22        1996          2011        0.00000   49722.7868
#> 218 Claim#22        1996          2012     3010.16852   47441.9250
#> 219 Claim#22        1996          2013        0.00000   47441.9250
#> 220 Claim#22        1996          2014        0.00000   47441.9250
#> 221 Claim#22        1996          2015        0.00000   47441.9250
#> 222 Claim#22        1996          2016      160.70385   47256.0470
#> 223 Claim#22        1996          2017        0.00000   47256.0470
#> 224 Claim#22        1996          2018     9550.86714   98306.7661
#> 225 Claim#22        1996          2019        0.00000   98306.7661
#> 226 Claim#22        1996          2020     3193.17088   44643.2021
#> 227 Claim#22        1996          2021    -1539.31011   89286.4042
#> 228 Claim#22        1996          2022        0.00000   89286.4042
#> 229 Claim#22        1996          2023        0.00000   89286.4042
#> 230 Claim#32        2000          2000        0.00000    4546.2556
#> 231 Claim#32        2000          2001     1448.18234  227309.2255
#> 232 Claim#32        2000          2002        0.00000  227309.2255
#> 233 Claim#32        2000          2003        0.00000  227309.2255
#> 234 Claim#32        2000          2004        0.00000  355662.4794
#> 235 Claim#32        2000          2005        0.00000  355662.4794
#> 236 Claim#32        2000          2006   126529.83392  168939.6777
#> 237 Claim#32        2000          2007        0.00000  168939.6777
#> 238 Claim#32        2000          2008    43106.35880   88559.9574
#> 239 Claim#32        2000          2009        0.00000   88559.9574
#> 240 Claim#32        2000          2010    17030.01717  168429.3021
#> 241 Claim#32        2000          2011        0.00000  168429.3021
#> 242 Claim#32        2000          2012        0.00000  168429.3021
#> 243 Claim#32        2000          2013    36680.40487  143570.2731
#> 244 Claim#32        2000          2014        0.00000  143570.2731
#> 245 Claim#32        2000          2015    30590.72319  133373.4298
#> 246 Claim#32        2000          2016        0.00000  133373.4298
#> 247 Claim#32        2000          2017        0.00000  133373.4298
#> 248 Claim#32        2000          2018    30178.79542  106025.6526
#> 249 Claim#32        2000          2019        0.00000  106025.6526
#> 250 Claim#32        2000          2020    16986.11182   85263.8553
#> 251 Claim#32        2000          2021        0.00000   85263.8553
#> 252 Claim#32        2000          2022        0.00000  167161.3653
#> 253 Claim#32        2000          2023    19131.28061  145280.1204
#> 254 Claim#33        2000          2000     4924.90251   27953.7757
#> 255 Claim#33        2000          2001     -276.08858  216136.4112
#> 256 Claim#33        2000          2002      195.23654  213251.9865
#> 257 Claim#33        2000          2003     -182.32709  212977.4885
#> 258 Claim#33        2000          2004      642.47191  212227.5600
#> 259 Claim#33        2000          2005     4770.92726  207065.9001
#> 260 Claim#33        2000          2006      802.25160  205865.7950
#> 261 Claim#33        2000          2007      778.53974  205042.3010
#> 262 Claim#33        2000          2008        0.00000  205042.3010
#> 263 Claim#33        2000          2009        0.00000  205042.3010
#> 264 Claim#33        2000          2010        0.00000  205042.3010
#> 265 Claim#33        2000          2011        0.00000  205042.3010
#> 266 Claim#33        2000          2012        0.00000  205042.3010
#> 267 Claim#33        2000          2013        0.00000  205042.3010
#> 268 Claim#33        2000          2014        0.00000   54899.5949
#> 269 Claim#33        2000          2015        0.00000   54899.5949
#> 270 Claim#33        2000          2016        0.00000   54899.5949
#> 271 Claim#33        2000          2017        0.00000   54899.5949
#> 272 Claim#33        2000          2018        0.00000   54899.5949
#> 273 Claim#33        2000          2019        0.00000   54899.5949
#> 274 Claim#33        2000          2020        0.00000   54899.5949
#> 275 Claim#33        2000          2021        0.00000   54899.5949
#> 276 Claim#33        2000          2022        0.00000   54899.5949
#> 277 Claim#33        2000          2023        0.00000   54899.5949
#> 278 Claim#34        2002          2002    10856.16880   34429.2432
#> 279 Claim#34        2002          2003    16112.95537  194378.2511
#> 280 Claim#34        2002          2004    13390.92644  177933.8163
#> 281 Claim#34        2002          2005     1518.76166  176387.4122
#> 282 Claim#34        2002          2006    16420.64988  160045.4967
#> 283 Claim#34        2002          2007    36714.71715  124710.8102
#> 284 Claim#34        2002          2008    66224.17118   69041.1258
#> 285 Claim#34        2002          2009    14336.80657   55779.0957
#> 286 Claim#34        2002          2010     -509.02124   53580.5313
#> 287 Claim#34        2002          2011        0.00000   53580.5313
#> 288 Claim#34        2002          2012      100.58542   53373.7697
#> 289 Claim#34        2002          2013        0.00000   53373.7697
#> 290 Claim#34        2002          2014        0.00000   53373.7697
#> 291 Claim#34        2002          2015        0.00000   53373.7697
#> 292 Claim#34        2002          2016    16567.39382   40451.1732
#> 293 Claim#34        2002          2017        0.00000   40451.1732
#> 294 Claim#34        2002          2018        0.00000   40451.1732
#> 295 Claim#34        2002          2019        0.00000   40451.1732
#> 296 Claim#34        2002          2020        0.00000   40451.1732
#> 297 Claim#34        2002          2021        0.00000   40451.1732
#> 298 Claim#34        2002          2022        0.00000   40451.1732
#> 299 Claim#34        2002          2023        0.00000   40451.1732
#> 300 Claim#37        2003          2003     9188.94986  378890.6870
#> 301 Claim#37        2003          2004    82453.99168  308802.8943
#> 302 Claim#37        2003          2005    47734.76703  267282.3974
#> 303 Claim#37        2003          2006   118313.41142  229324.6223
#> 304 Claim#37        2003          2007      856.20373  228269.0010
#> 305 Claim#37        2003          2008        0.00000  228269.0010
#> 306 Claim#37        2003          2009        0.00000  309406.2364
#> 307 Claim#37        2003          2010      659.89887  308563.5595
#> 308 Claim#37        2003          2011   138046.78285  182003.6685
#> 309 Claim#37        2003          2012        0.00000  182003.6685
#> 310 Claim#37        2003          2013        0.00000  182003.6685
#> 311 Claim#37        2003          2014        0.00000  182003.6685
#> 312 Claim#37        2003          2015        0.00000    9100.1834
#> 313 Claim#37        2003          2016        0.00000    9100.1834
#> 314 Claim#37        2003          2017        0.00000    9100.1834
#> 315 Claim#37        2003          2018        0.00000    9100.1834
#> 316 Claim#37        2003          2019        0.00000    9100.1834
#> 317 Claim#39        2004          2004    22018.14978  214069.7330
#> 318 Claim#39        2004          2005    38067.22907  169316.5330
#> 319 Claim#39        2004          2006    25625.90038   87861.4628
#> 320 Claim#39        2004          2007     7650.65324       0.0000
#> 321 Claim#40        2004          2004        0.00000    5554.7808
#> 322 Claim#40        2004          2005        0.00000    5554.7808
#> 323 Claim#40        2004          2006     4481.08859   55547.8076
#> 324 Claim#40        2004          2007    11659.97878   98766.2238
#> 325 Claim#40        2004          2008    13567.04596  186085.1553
#> 326 Claim#40        2004          2009   114101.51871   69138.1342
#> 327 Claim#40        2004          2010     1930.61214   67071.7557
#> 328 Claim#40        2004          2011     1784.73043   65044.2607
#> 329 Claim#40        2004          2012     2064.34206   62964.5508
#> 330 Claim#40        2004          2013     1703.49188   60799.2973
#> 331 Claim#40        2004          2014     2206.43983   58634.0437
#> 332 Claim#40        2004          2015     1989.85468   56365.4713
#> 333 Claim#40        2004          2016     2225.78129   53998.0237
#> 334 Claim#40        2004          2017     4016.36697   49687.5139
#> 335 Claim#40        2004          2018     4118.12717   62440.1795
#> 336 Claim#40        2004          2019     4246.85887   58124.1149
#> 337 Claim#40        2004          2020     4224.54983   53512.5359
#> 338 Claim#40        2004          2021     3821.89455   48768.7531
#> 339 Claim#40        2004          2022     3952.18147   43573.9222
#> 340 Claim#40        2004          2023     4631.47926   38130.2370
#> 341 Claim#41        2005          2005        0.00000   11137.5470
#> 342 Claim#41        2005          2006   124406.73945  667091.1755
#> 343 Claim#41        2005          2007   110597.31166  526617.7498
#> 344 Claim#41        2005          2008    67554.30833  812643.3226
#> 345 Claim#41        2005          2009    40108.60149  769738.1502
#> 346 Claim#41        2005          2010        0.00000  769738.1502
#> 347 Claim#41        2005          2011    67504.31301  694421.6021
#> 348 Claim#41        2005          2012     4857.09076       0.0000
#> 349 Claim#43        2006          2007   157068.32116  530498.1165
#> 350 Claim#43        2006          2008   158661.76773  391067.3085
#> 351 Claim#43        2006          2009    64912.80964  335066.4559
#> 352 Claim#43        2006          2010    32140.96035  301030.8779
#> 353 Claim#43        2006          2011    24380.25210  276512.5763
#> 354 Claim#43        2006          2012    59815.29072  228138.1347
#> 355 Claim#43        2006          2013    31598.12503  198664.3792
#> 356 Claim#43        2006          2014    29748.03120  307779.9847
#> 357 Claim#43        2006          2015    41539.87579  276862.3825
#> 358 Claim#43        2006          2016    24742.13854  257536.1918
#> 359 Claim#43        2006          2017    47703.16051  219210.5087
#> 360 Claim#43        2006          2018    76169.25544  154734.6267
#> 361 Claim#43        2006          2019    65647.42382  105533.0808
#> 362 Claim#43        2006          2020    34324.94020   79178.0979
#> 363 Claim#43        2006          2021    41322.95106  219008.1152
#> 364 Claim#43        2006          2022    30500.22110  194535.2327
#> 365 Claim#43        2006          2023   112133.08492  240134.3333
#> 366 Claim#44        2006          2006    18503.07736  193098.5685
#> 367 Claim#44        2006          2007    91188.28066 1152198.2091
#> 368 Claim#44        2006          2008    28168.99576 1115369.0781
#> 369 Claim#44        2006          2009    19079.97029 1089706.7844
#> 370 Claim#44        2006          2010    50403.02922 1028570.8909
#> 371 Claim#44        2006          2011    69888.45496  935281.1786
#> 372 Claim#44        2006          2012      540.60046  934577.4514
#> 373 Claim#44        2006          2013    11878.79053  917686.8918
#> 374 Claim#44        2006          2014        0.00000  917686.8918
#> 375 Claim#44        2006          2015        0.00000  917686.8918
#> 376 Claim#44        2006          2016    10193.66856  906220.6666
#> 377 Claim#44        2006          2017        0.00000  906220.6666
#> 378 Claim#44        2006          2018        0.00000  758418.0515
#> 379 Claim#44        2006          2019        0.00000  758328.5665
#> 380 Claim#44        2006          2020        0.00000  758328.5665
#> 381 Claim#44        2006          2021        0.00000  751913.2693
#> 382 Claim#44        2006          2022      195.16353  514406.9682
#> 383 Claim#44        2006          2023        0.00000  514406.9682
#> 384 Claim#45        2006          2006    21419.64474  409778.6991
#> 385 Claim#45        2006          2007    54425.13879  359521.1660
#> 386 Claim#45        2006          2008    41437.47738  322585.6292
#> 387 Claim#45        2006          2009    16047.04291  359943.6035
#> 388 Claim#45        2006          2010     9681.84566  348902.9488
#> 389 Claim#45        2006          2011     6515.84240  341489.7076
#> 390 Claim#45        2006          2012    89984.36069  327030.1730
#> 391 Claim#45        2006          2013    19660.75457  305233.4718
#> 392 Claim#45        2006          2014        0.00000  305233.4718
#> 393 Claim#45        2006          2015        0.00000  305233.4718
#> 394 Claim#45        2006          2016      149.19151  305104.5925
#> 395 Claim#45        2006          2017        0.00000  305104.5925
#> 396 Claim#45        2006          2018    14825.69038  291679.6718
#> 397 Claim#45        2006          2019        0.00000  291679.6718
#> 398 Claim#45        2006          2020        0.00000  291679.6718
#> 399 Claim#45        2006          2021        0.00000  268498.4138
#> 400 Claim#45        2006          2022        0.00000  268498.4138
#> 401 Claim#45        2006          2023      179.72730  268343.5797
#> 402 Claim#46        2008          2008     2378.65805   63354.2429
#> 403 Claim#46        2008          2009    21081.01239   43423.0115
#> 404 Claim#46        2008          2010    18553.44607   29510.8537
#> 405 Claim#46        2008          2011     2353.34572   27781.5462
#> 406 Claim#46        2008          2012     3002.41631  219806.6754
#> 407 Claim#46        2008          2013     1433.22268  218850.7804
#> 408 Claim#46        2008          2014     2661.67244  216318.8346
#> 409 Claim#46        2008          2015     6065.19502  212294.9513
#> 410 Claim#46        2008          2016       93.94696  212230.7973
#> 411 Claim#46        2008          2017        0.00000  212230.7973
#> 412 Claim#46        2008          2018       75.88408  212165.9304
#> 413 Claim#46        2008          2019        0.00000  212165.9304
#> 414 Claim#46        2008          2020     3855.48385  239538.3165
#> 415 Claim#46        2008          2021        0.00000  239538.3165
#> 416 Claim#46        2008          2022     2328.60007  237215.9406
#> 417 Claim#46        2008          2023     1137.32523  236463.9128
#> 418 Claim#49        2010          2010    24293.96837   74594.1259
#> 419 Claim#49        2010          2011    98242.14941  507713.5325
#> 420 Claim#49        2010          2012    38921.63647  456197.5893
#> 421 Claim#49        2010          2013    39055.00343  416723.9409
#> 422 Claim#49        2010          2014    38888.35625  365463.1127
#> 423 Claim#49        2010          2015    40943.98179  156059.0621
#> 424 Claim#49        2010          2016    10155.60477  144547.5362
#> 425 Claim#49        2010          2017        0.00000  144547.5362
#> 426 Claim#49        2010          2018   100184.17196       0.0000
#> 427 Claim#51        2010          2010     5675.37124   22941.4950
#> 428 Claim#51        2010          2011     9513.83914  300057.5478
#> 429 Claim#51        2010          2012    35879.56765  261114.3602
#> 430 Claim#51        2010          2013    22482.03745   90618.9051
#> 431 Claim#51        2010          2014        0.00000   90618.9051
#> 432 Claim#51        2010          2015       23.28973   90591.3753
#> 433 Claim#51        2010          2016     1018.86462   89063.4717
#> 434 Claim#51        2010          2017      857.02381  160590.4647
#> 435 Claim#51        2010          2018       12.72103  160590.4647
#> 436 Claim#51        2010          2019        0.00000  160590.4647
#> 437 Claim#51        2010          2020        0.00000  160590.4647
#> 438 Claim#51        2010          2021        0.00000  160590.4647
#> 439 Claim#51        2010          2022        0.00000  160590.4647
#> 440 Claim#51        2010          2023        0.00000  160590.4647
#> 441 Claim#52        2012          2012    37709.25547  363958.8969
#> 442 Claim#52        2012          2013    71020.55918  316992.6164
#> 443 Claim#52        2012          2014    13617.47030  304245.7891
#> 444 Claim#52        2012          2015    31464.01594  271718.7262
#> 445 Claim#52        2012          2016      235.02938  271486.5408
#> 446 Claim#52        2012          2017       65.67169  271421.7098
#> 447 Claim#52        2012          2018        0.00000  271421.7098
#> 448 Claim#52        2012          2019      162.23295  222272.2835
#> 449 Claim#52        2012          2020        0.00000  222272.2835
#> 450 Claim#52        2012          2021        0.00000  214846.8740
#> 451 Claim#52        2012          2022        0.00000  214846.8740
#> 452 Claim#52        2012          2023        0.00000  214846.8740
#> 453 Claim#53        2012          2012        0.00000    8960.4620
#> 454 Claim#53        2012          2013       23.75150     151.4318
#> 455 Claim#53        2012          2014        0.00000     151.4318
#> 456 Claim#53        2012          2015       63.82755  224011.5501
#> 457 Claim#53        2012          2016       77.60261  223950.6190
#> 458 Claim#53        2012          2017        0.00000  223950.6190
#> 459 Claim#53        2012          2018        0.00000  223950.6190
#> 460 Claim#53        2012          2019        0.00000  223950.6190
#> 461 Claim#53        2012          2020        0.00000  223950.6190
#> 462 Claim#53        2012          2021        0.00000  223950.6190
#> 463 Claim#53        2012          2022        0.00000  223950.6190
#> 464 Claim#53        2012          2023        0.00000  223950.6190
#> 465 Claim#57        2015          2015     8526.03162  198671.9184
#> 466 Claim#57        2015          2016    35277.46183  170492.3478
#> 467 Claim#57        2015          2017   103567.64623   63863.5079
#> 468 Claim#57        2015          2018    28417.93614   39107.6023
#> 469 Claim#57        2015          2019   -57856.95214  116157.7524
#> 470 Claim#57        2015          2020     3815.28699  104532.7952
#> 471 Claim#57        2015          2021     6848.61792   86174.1534
#> 472 Claim#57        2015          2022     -741.59929   81674.9955
#> 473 Claim#57        2015          2023     5057.98698   74848.7387
#> 474 Claim#60        2015          2015        0.00000    3307.3252
#> 475 Claim#60        2015          2016    20255.90715  250254.2761
#> 476 Claim#60        2015          2017    11160.25018  235922.5334
#> 477 Claim#60        2015          2018    36301.07986  296524.8586
#> 478 Claim#60        2015          2019    21776.99964  266134.9494
#> 479 Claim#60        2015          2020    22294.39700  160960.9045
#> 480 Claim#60        2015          2021    58573.62787   85119.5271
#> 481 Claim#60        2015          2022    30843.70932   47952.9086
#> 482 Claim#60        2015          2023        0.00000   27561.0436
#> 483 Claim#63        2017          2017    13477.06529  258306.4988
#> 484 Claim#63        2017          2018    44325.76893  203698.0016
#> 485 Claim#63        2017          2019    21150.19715  171116.1074
#> 486 Claim#63        2017          2020    14002.02242   54067.9881
#> 487 Claim#63        2017          2021      492.11910   53514.9618
#> 488 Claim#63        2017          2022     1421.11586   51665.8514
#> 489 Claim#63        2017          2023     2357.22598   48945.0988
#> 490 Claim#65        2019          2019    32005.78194   87085.0882
#> 491 Claim#65        2019          2020     3761.23800  122387.0858
#> 492 Claim#65        2019          2021        0.00000  122387.0858
#> 493 Claim#65        2019          2022   267856.20615   89050.8865
#> 494 Claim#65        2019          2023    48796.81794   26184.3000
#> 495 Claim#68        2019          2019    13386.85862   53779.8443
#> 496 Claim#68        2019          2020       86.22699  113116.5657
#> 497 Claim#68        2019          2021    36235.60978  179474.3466
#> 498 Claim#68        2019          2022    75743.97319  294513.3587
#> 499 Claim#68        2019          2023     3524.88778  289074.1593
#>     An_payment_cal An_reserve
#> 1            0.000        0.0
#> 2            0.000        0.0
#> 3            0.000        0.0
#> 4            0.000        0.0
#> 5            0.000        0.0
#> 6            0.000        0.0
#> 7            0.000        0.0
#> 8            0.000        0.0
#> 9            0.000        0.0
#> 10           0.000        0.0
#> 11           0.000        0.0
#> 12           0.000        0.0
#> 13           0.000        0.0
#> 14           0.000        0.0
#> 15           0.000        0.0
#> 16           0.000        0.0
#> 17           0.000        0.0
#> 18           0.000        0.0
#> 19           0.000        0.0
#> 20           0.000        0.0
#> 21           0.000        0.0
#> 22           0.000        0.0
#> 23           0.000        0.0
#> 24           0.000        0.0
#> 25           0.000        0.0
#> 26           0.000        0.0
#> 27           0.000        0.0
#> 28           0.000        0.0
#> 29           0.000        0.0
#> 30           0.000        0.0
#> 31           0.000        0.0
#> 32           0.000        0.0
#> 33           0.000        0.0
#> 34           0.000        0.0
#> 35           0.000        0.0
#> 36           0.000        0.0
#> 37           0.000        0.0
#> 38           0.000        0.0
#> 39           0.000        0.0
#> 40           0.000        0.0
#> 41           0.000        0.0
#> 42           0.000        0.0
#> 43      263568.208   239848.7
#> 44           0.000        0.0
#> 45           0.000        0.0
#> 46           0.000        0.0
#> 47           0.000        0.0
#> 48           0.000        0.0
#> 49           0.000        0.0
#> 50           0.000        0.0
#> 51           0.000        0.0
#> 52           0.000        0.0
#> 53           0.000        0.0
#> 54           0.000        0.0
#> 55           0.000        0.0
#> 56           0.000        0.0
#> 57           0.000        0.0
#> 58           0.000        0.0
#> 59           0.000        0.0
#> 60           0.000        0.0
#> 61           0.000        0.0
#> 62           0.000        0.0
#> 63           0.000        0.0
#> 64           0.000        0.0
#> 65           0.000        0.0
#> 66           0.000        0.0
#> 67           0.000        0.0
#> 68           0.000        0.0
#> 69           0.000        0.0
#> 70           0.000        0.0
#> 71           0.000        0.0
#> 72           0.000        0.0
#> 73           0.000        0.0
#> 74           0.000        0.0
#> 75           0.000        0.0
#> 76           0.000        0.0
#> 77           0.000        0.0
#> 78           0.000        0.0
#> 79           0.000        0.0
#> 80           0.000        0.0
#> 81           0.000        0.0
#> 82           0.000        0.0
#> 83           0.000        0.0
#> 84           0.000        0.0
#> 85           0.000        0.0
#> 86           0.000        0.0
#> 87           0.000        0.0
#> 88           0.000        0.0
#> 89           0.000        0.0
#> 90           0.000        0.0
#> 91           0.000        0.0
#> 92           0.000        0.0
#> 93           0.000        0.0
#> 94           0.000        0.0
#> 95           0.000        0.0
#> 96           0.000        0.0
#> 97           0.000        0.0
#> 98           0.000        0.0
#> 99           0.000        0.0
#> 100          0.000        0.0
#> 101          0.000        0.0
#> 102          0.000        0.0
#> 103          0.000        0.0
#> 104          0.000        0.0
#> 105          0.000        0.0
#> 106          0.000        0.0
#> 107          0.000        0.0
#> 108          0.000        0.0
#> 109          0.000        0.0
#> 110          0.000        0.0
#> 111          0.000        0.0
#> 112          0.000        0.0
#> 113          0.000        0.0
#> 114          0.000        0.0
#> 115          0.000        0.0
#> 116          0.000        0.0
#> 117          0.000        0.0
#> 118          0.000        0.0
#> 119          0.000        0.0
#> 120          0.000        0.0
#> 121      56883.938   443502.9
#> 122      21083.744   438185.8
#> 123      12298.850   432726.8
#> 124      21083.744   414962.6
#> 125      21522.990   431330.0
#> 126      22137.925   426795.8
#> 127      22137.925   415417.6
#> 128      22137.925   410438.9
#> 129      22137.937   405314.2
#> 130      23982.765   439140.5
#> 131      20425.339   485858.8
#> 132      13745.405   480742.1
#> 133      13745.405   472350.5
#> 134      13745.405   260020.7
#> 135      13745.405   257570.6
#> 136      15979.033   281016.8
#> 137      22679.918   269976.9
#> 138      22679.918   287163.3
#> 139      22679.918   283737.2
#> 140      14489.948   280252.8
#> 141      13745.405   288068.2
#> 142      13745.405   283830.0
#> 143      13745.405   324866.9
#> 144      13745.405   345009.7
#> 145      13745.405   337912.5
#> 146      13745.405   330785.5
#> 147      13745.405   341439.3
#> 148      13745.405   369003.4
#> 149      13745.405   359247.6
#> 150      13745.405   349522.7
#> 151          0.000        0.0
#> 152          0.000        0.0
#> 153          0.000        0.0
#> 154          0.000        0.0
#> 155          0.000        0.0
#> 156          0.000        0.0
#> 157          0.000        0.0
#> 158          0.000        0.0
#> 159          0.000        0.0
#> 160          0.000        0.0
#> 161          0.000        0.0
#> 162          0.000        0.0
#> 163          0.000        0.0
#> 164          0.000        0.0
#> 165          0.000        0.0
#> 166          0.000        0.0
#> 167          0.000        0.0
#> 168          0.000        0.0
#> 169          0.000        0.0
#> 170          0.000        0.0
#> 171          0.000        0.0
#> 172          0.000        0.0
#> 173          0.000        0.0
#> 174          0.000        0.0
#> 175          0.000        0.0
#> 176          0.000        0.0
#> 177          0.000        0.0
#> 178          0.000        0.0
#> 179          0.000        0.0
#> 180          0.000        0.0
#> 181          0.000        0.0
#> 182          0.000        0.0
#> 183          0.000        0.0
#> 184          0.000        0.0
#> 185          0.000        0.0
#> 186          0.000        0.0
#> 187          0.000        0.0
#> 188          0.000        0.0
#> 189          0.000        0.0
#> 190          0.000        0.0
#> 191          0.000        0.0
#> 192          0.000        0.0
#> 193          0.000        0.0
#> 194          0.000        0.0
#> 195          0.000        0.0
#> 196          0.000        0.0
#> 197          0.000        0.0
#> 198          0.000        0.0
#> 199          0.000        0.0
#> 200          0.000        0.0
#> 201          0.000        0.0
#> 202          0.000        0.0
#> 203          0.000        0.0
#> 204          0.000        0.0
#> 205          0.000        0.0
#> 206          0.000        0.0
#> 207          0.000        0.0
#> 208          0.000        0.0
#> 209          0.000        0.0
#> 210          0.000   307501.8
#> 211       7381.126   312272.5
#> 212       9207.484   353022.3
#> 213       6984.490   362587.6
#> 214       5438.248   421803.1
#> 215          0.000   427796.2
#> 216          0.000   434121.3
#> 217       3363.009   440802.2
#> 218          0.000   447860.9
#> 219       5190.752   455323.7
#> 220       3225.921   463215.7
#> 221       4828.718   471570.0
#> 222       7439.855   493810.6
#> 223       6252.488   457198.5
#> 224      12474.908   367773.8
#> 225          0.000   376080.1
#> 226          0.000   388520.7
#> 227          0.000        0.0
#> 228          0.000        0.0
#> 229          0.000        0.0
#> 230          0.000        0.0
#> 231          0.000        0.0
#> 232          0.000        0.0
#> 233          0.000        0.0
#> 234          0.000        0.0
#> 235          0.000        0.0
#> 236          0.000        0.0
#> 237          0.000        0.0
#> 238          0.000        0.0
#> 239          0.000        0.0
#> 240          0.000        0.0
#> 241          0.000        0.0
#> 242          0.000        0.0
#> 243          0.000        0.0
#> 244          0.000        0.0
#> 245          0.000        0.0
#> 246          0.000        0.0
#> 247          0.000        0.0
#> 248          0.000        0.0
#> 249          0.000        0.0
#> 250          0.000        0.0
#> 251          0.000        0.0
#> 252          0.000        0.0
#> 253          0.000        0.0
#> 254          0.000        0.0
#> 255          0.000        0.0
#> 256          0.000        0.0
#> 257          0.000        0.0
#> 258          0.000        0.0
#> 259          0.000        0.0
#> 260          0.000        0.0
#> 261          0.000        0.0
#> 262          0.000        0.0
#> 263          0.000        0.0
#> 264          0.000        0.0
#> 265          0.000        0.0
#> 266          0.000        0.0
#> 267          0.000        0.0
#> 268          0.000        0.0
#> 269          0.000        0.0
#> 270          0.000        0.0
#> 271          0.000        0.0
#> 272          0.000        0.0
#> 273          0.000        0.0
#> 274          0.000        0.0
#> 275          0.000        0.0
#> 276          0.000        0.0
#> 277          0.000        0.0
#> 278          0.000        0.0
#> 279          0.000        0.0
#> 280          0.000        0.0
#> 281          0.000        0.0
#> 282          0.000        0.0
#> 283          0.000        0.0
#> 284          0.000        0.0
#> 285          0.000        0.0
#> 286          0.000        0.0
#> 287          0.000        0.0
#> 288          0.000        0.0
#> 289          0.000        0.0
#> 290          0.000        0.0
#> 291          0.000        0.0
#> 292          0.000        0.0
#> 293          0.000        0.0
#> 294          0.000        0.0
#> 295          0.000        0.0
#> 296          0.000        0.0
#> 297          0.000        0.0
#> 298          0.000        0.0
#> 299          0.000        0.0
#> 300          0.000        0.0
#> 301          0.000        0.0
#> 302          0.000        0.0
#> 303          0.000        0.0
#> 304          0.000        0.0
#> 305          0.000        0.0
#> 306          0.000        0.0
#> 307          0.000        0.0
#> 308          0.000        0.0
#> 309          0.000        0.0
#> 310          0.000        0.0
#> 311          0.000        0.0
#> 312          0.000        0.0
#> 313          0.000        0.0
#> 314          0.000        0.0
#> 315          0.000        0.0
#> 316          0.000        0.0
#> 317          0.000        0.0
#> 318          0.000        0.0
#> 319          0.000        0.0
#> 320          0.000        0.0
#> 321          0.000        0.0
#> 322          0.000        0.0
#> 323          0.000        0.0
#> 324          0.000        0.0
#> 325          0.000        0.0
#> 326          0.000        0.0
#> 327          0.000        0.0
#> 328          0.000        0.0
#> 329          0.000        0.0
#> 330          0.000        0.0
#> 331          0.000        0.0
#> 332          0.000        0.0
#> 333          0.000        0.0
#> 334          0.000        0.0
#> 335          0.000        0.0
#> 336          0.000        0.0
#> 337          0.000        0.0
#> 338          0.000        0.0
#> 339          0.000        0.0
#> 340          0.000        0.0
#> 341          0.000        0.0
#> 342          0.000        0.0
#> 343          0.000        0.0
#> 344          0.000        0.0
#> 345          0.000        0.0
#> 346          0.000        0.0
#> 347          0.000        0.0
#> 348          0.000        0.0
#> 349          0.000        0.0
#> 350          0.000        0.0
#> 351       3605.612   184681.8
#> 352      14422.448   178173.8
#> 353      14422.448   171647.8
#> 354      14422.448   165127.5
#> 355      14422.448   158641.8
#> 356      14422.448   152217.8
#> 357      14422.448   145879.9
#> 358      14422.448   144695.2
#> 359      14422.448   143153.3
#> 360      14422.448   136447.1
#> 361      14422.448   129974.9
#> 362      14422.448   126573.1
#> 363      14422.448   125464.5
#> 364      14422.448   119125.5
#> 365      14422.448   113123.6
#> 366          0.000        0.0
#> 367          0.000        0.0
#> 368          0.000        0.0
#> 369          0.000        0.0
#> 370          0.000        0.0
#> 371          0.000        0.0
#> 372          0.000        0.0
#> 373          0.000        0.0
#> 374          0.000        0.0
#> 375          0.000        0.0
#> 376          0.000        0.0
#> 377          0.000        0.0
#> 378          0.000        0.0
#> 379          0.000        0.0
#> 380          0.000        0.0
#> 381          0.000        0.0
#> 382          0.000        0.0
#> 383          0.000        0.0
#> 384          0.000        0.0
#> 385          0.000        0.0
#> 386          0.000        0.0
#> 387          0.000        0.0
#> 388          0.000        0.0
#> 389          0.000        0.0
#> 390          0.000        0.0
#> 391          0.000        0.0
#> 392          0.000        0.0
#> 393          0.000        0.0
#> 394          0.000        0.0
#> 395          0.000        0.0
#> 396          0.000        0.0
#> 397          0.000        0.0
#> 398          0.000        0.0
#> 399          0.000        0.0
#> 400          0.000        0.0
#> 401          0.000        0.0
#> 402          0.000        0.0
#> 403          0.000        0.0
#> 404          0.000        0.0
#> 405          0.000        0.0
#> 406          0.000        0.0
#> 407          0.000        0.0
#> 408          0.000        0.0
#> 409          0.000        0.0
#> 410          0.000        0.0
#> 411          0.000        0.0
#> 412          0.000        0.0
#> 413          0.000        0.0
#> 414          0.000        0.0
#> 415          0.000        0.0
#> 416          0.000        0.0
#> 417          0.000        0.0
#> 418          0.000        0.0
#> 419          0.000        0.0
#> 420          0.000        0.0
#> 421          0.000        0.0
#> 422          0.000        0.0
#> 423          0.000        0.0
#> 424          0.000        0.0
#> 425          0.000        0.0
#> 426          0.000        0.0
#> 427          0.000        0.0
#> 428          0.000        0.0
#> 429          0.000        0.0
#> 430          0.000   121862.2
#> 431          0.000   124217.9
#> 432          0.000   126628.6
#> 433          0.000   129098.9
#> 434          0.000   148915.8
#> 435          0.000   151104.3
#> 436          0.000   153339.8
#> 437          0.000   168019.3
#> 438          0.000   194934.4
#> 439          0.000   195934.1
#> 440          0.000   196963.7
#> 441          0.000        0.0
#> 442          0.000        0.0
#> 443          0.000        0.0
#> 444          0.000        0.0
#> 445          0.000        0.0
#> 446          0.000        0.0
#> 447          0.000        0.0
#> 448          0.000        0.0
#> 449          0.000        0.0
#> 450          0.000        0.0
#> 451          0.000        0.0
#> 452          0.000        0.0
#> 453          0.000        0.0
#> 454          0.000        0.0
#> 455          0.000        0.0
#> 456          0.000        0.0
#> 457          0.000        0.0
#> 458          0.000        0.0
#> 459          0.000        0.0
#> 460          0.000        0.0
#> 461          0.000        0.0
#> 462          0.000        0.0
#> 463          0.000        0.0
#> 464          0.000        0.0
#> 465          0.000        0.0
#> 466          0.000        0.0
#> 467          0.000        0.0
#> 468          0.000        0.0
#> 469          0.000        0.0
#> 470          0.000        0.0
#> 471          0.000        0.0
#> 472          0.000        0.0
#> 473          0.000        0.0
#> 474          0.000        0.0
#> 475          0.000        0.0
#> 476          0.000        0.0
#> 477          0.000        0.0
#> 478          0.000        0.0
#> 479          0.000        0.0
#> 480          0.000        0.0
#> 481          0.000        0.0
#> 482          0.000        0.0
#> 483          0.000        0.0
#> 484          0.000        0.0
#> 485          0.000        0.0
#> 486          0.000        0.0
#> 487          0.000        0.0
#> 488          0.000        0.0
#> 489          0.000        0.0
#> 490          0.000        0.0
#> 491          0.000        0.0
#> 492          0.000        0.0
#> 493          0.000        0.0
#> 494          0.000        0.0
#> 495          0.000        0.0
#> 496          0.000        0.0
#> 497          0.000        0.0
#> 498          0.000        0.0
#> 499          0.000        0.0