Rolling out active annuities corresponding to large claims

Rolling out active annuities corresponding to large claims

Usage

roll_out_active_annuities(
  payments_active_annuities,
  probs_active_annuities,
  active_annuities,
  large_claims_list
)

Arguments

payments_active_annuities

matrix of agreed future payments corresponding to active annuities, see generate_annuity_payments().

probs_active_annuities

matrix of mortality probabilities corresponding to active annuities, see generate_annuity_probabilities().

active_annuities

Dataframe of active annuities, see active_annuities_xmpl.

large_claims_list

Dataframe of large claims generated with generate_claims_list().

Value

Numeric matrix with one row per large claim and 130 future annuity payments.

Examples

# this example uses data provided with this package
print(minimal_active_annuities_xmpl)
#>   Claim_id Annuity_id Origin_year Calendar_year Entering_year Annuity_start
#> 1 Claim#43          1        2006          2023          2009          2009
#> 2 Claim#51          1        2010          2023          2013          2036
#> 3 Claim#13          1        1990          2023          1992          1992
#>   Annuity_end Birth_year Gender Annual_payment Dynamic
#> 1           0       1939      m       12168.21       0
#> 2        2051       1986      m       12893.00       0
#> 3           0       1958      w       11807.36       0

# Create payments matrix
payments <- generate_annuity_payments(annuities = minimal_active_annuities_xmpl,
                                      last_orig_year = 2023)
print(payments[,1:10])
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 12168.21 12168.21 12168.21 12168.21 12168.21 12168.21 12168.21 12168.21
#> [2,]     0.00     0.00     0.00     0.00     0.00     0.00     0.00     0.00
#> [3,] 11807.36 11807.36 11807.36 11807.36 11807.36 11807.36 11807.36 11807.36
#>          [,9]    [,10]
#> [1,] 12168.21 12168.21
#> [2,]     0.00     0.00
#> [3,] 11807.36 11807.36

# Create probability matrix
probs <- generate_annuity_probabilities(annuities = minimal_active_annuities_xmpl,
                                        mortality = mortality_xmpl,
                                        age_shift = age_shift_xmpl)
print(probs[,1:10])
#>      [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]      [,8]
#> [1,]    0 0.2868083 0.4937320 0.6423048 0.7484672 0.8239585 0.8773790 0.9149971
#> [2,]    0 0.1136917 0.2174082 0.3115931 0.3967347 0.4733544 0.5419961 0.6032160
#> [3,]    0 0.1839400 0.3364235 0.4623482 0.5659430 0.6508416 0.7201519 0.7765189
#>           [,9]     [,10]
#> [1,] 0.9413576 0.9597385
#> [2,] 0.6575737 0.7056247
#> [3,] 0.8221835 0.8590354

# Create large claims list
extended_claims_data <- prepare_data(claims_data = minimal_claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = minimal_pool_of_annuities_xmpl)

large_claims_list <- generate_claims_list(extended_claims_data = extended_claims_data,
                                          first_orig_year = 1989,
                                          last_orig_year = 2023)

roll_out_active_annuities(payments_active_annuities = payments,
                          probs_active_annuities = probs,
                          active_annuities = minimal_active_annuities_xmpl,
                          large_claims_list = large_claims_list)
#>              [,1]     [,2]     [,3]     [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> Claim#10     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#13 11807.36 11807.36     0.00     0.00    0    0    0    0    0     0
#> Claim#16     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#22     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#32     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#37     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#41     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#43 12168.21 12168.21 12168.21 12168.21    0    0    0    0    0     0
#> Claim#44     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#45     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#49     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#51     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#52     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#60     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#> Claim#68     0.00     0.00     0.00     0.00    0    0    0    0    0     0
#>          [,11] [,12] [,13] [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,22] [,23] [,24] [,25] [,26] [,27] [,28] [,29] [,30] [,31] [,32]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,33] [,34] [,35] [,36] [,37] [,38] [,39] [,40] [,41] [,42] [,43]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,44] [,45] [,46] [,47] [,48] [,49] [,50] [,51] [,52] [,53] [,54]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,55] [,56] [,57] [,58] [,59] [,60] [,61] [,62] [,63] [,64] [,65]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,66] [,67] [,68] [,69] [,70] [,71] [,72] [,73] [,74] [,75] [,76]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,77] [,78] [,79] [,80] [,81] [,82] [,83] [,84] [,85] [,86] [,87]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,88] [,89] [,90] [,91] [,92] [,93] [,94] [,95] [,96] [,97] [,98]
#> Claim#10     0     0     0     0     0     0     0     0     0     0     0
#> Claim#13     0     0     0     0     0     0     0     0     0     0     0
#> Claim#16     0     0     0     0     0     0     0     0     0     0     0
#> Claim#22     0     0     0     0     0     0     0     0     0     0     0
#> Claim#32     0     0     0     0     0     0     0     0     0     0     0
#> Claim#37     0     0     0     0     0     0     0     0     0     0     0
#> Claim#41     0     0     0     0     0     0     0     0     0     0     0
#> Claim#43     0     0     0     0     0     0     0     0     0     0     0
#> Claim#44     0     0     0     0     0     0     0     0     0     0     0
#> Claim#45     0     0     0     0     0     0     0     0     0     0     0
#> Claim#49     0     0     0     0     0     0     0     0     0     0     0
#> Claim#51     0     0     0     0     0     0     0     0     0     0     0
#> Claim#52     0     0     0     0     0     0     0     0     0     0     0
#> Claim#60     0     0     0     0     0     0     0     0     0     0     0
#> Claim#68     0     0     0     0     0     0     0     0     0     0     0
#>          [,99] [,100] [,101] [,102] [,103] [,104] [,105] [,106] [,107] [,108]
#> Claim#10     0      0      0      0      0      0      0      0      0      0
#> Claim#13     0      0      0      0      0      0      0      0      0      0
#> Claim#16     0      0      0      0      0      0      0      0      0      0
#> Claim#22     0      0      0      0      0      0      0      0      0      0
#> Claim#32     0      0      0      0      0      0      0      0      0      0
#> Claim#37     0      0      0      0      0      0      0      0      0      0
#> Claim#41     0      0      0      0      0      0      0      0      0      0
#> Claim#43     0      0      0      0      0      0      0      0      0      0
#> Claim#44     0      0      0      0      0      0      0      0      0      0
#> Claim#45     0      0      0      0      0      0      0      0      0      0
#> Claim#49     0      0      0      0      0      0      0      0      0      0
#> Claim#51     0      0      0      0      0      0      0      0      0      0
#> Claim#52     0      0      0      0      0      0      0      0      0      0
#> Claim#60     0      0      0      0      0      0      0      0      0      0
#> Claim#68     0      0      0      0      0      0      0      0      0      0
#>          [,109] [,110] [,111] [,112] [,113] [,114] [,115] [,116] [,117] [,118]
#> Claim#10      0      0      0      0      0      0      0      0      0      0
#> Claim#13      0      0      0      0      0      0      0      0      0      0
#> Claim#16      0      0      0      0      0      0      0      0      0      0
#> Claim#22      0      0      0      0      0      0      0      0      0      0
#> Claim#32      0      0      0      0      0      0      0      0      0      0
#> Claim#37      0      0      0      0      0      0      0      0      0      0
#> Claim#41      0      0      0      0      0      0      0      0      0      0
#> Claim#43      0      0      0      0      0      0      0      0      0      0
#> Claim#44      0      0      0      0      0      0      0      0      0      0
#> Claim#45      0      0      0      0      0      0      0      0      0      0
#> Claim#49      0      0      0      0      0      0      0      0      0      0
#> Claim#51      0      0      0      0      0      0      0      0      0      0
#> Claim#52      0      0      0      0      0      0      0      0      0      0
#> Claim#60      0      0      0      0      0      0      0      0      0      0
#> Claim#68      0      0      0      0      0      0      0      0      0      0
#>          [,119] [,120] [,121] [,122] [,123] [,124] [,125] [,126] [,127] [,128]
#> Claim#10      0      0      0      0      0      0      0      0      0      0
#> Claim#13      0      0      0      0      0      0      0      0      0      0
#> Claim#16      0      0      0      0      0      0      0      0      0      0
#> Claim#22      0      0      0      0      0      0      0      0      0      0
#> Claim#32      0      0      0      0      0      0      0      0      0      0
#> Claim#37      0      0      0      0      0      0      0      0      0      0
#> Claim#41      0      0      0      0      0      0      0      0      0      0
#> Claim#43      0      0      0      0      0      0      0      0      0      0
#> Claim#44      0      0      0      0      0      0      0      0      0      0
#> Claim#45      0      0      0      0      0      0      0      0      0      0
#> Claim#49      0      0      0      0      0      0      0      0      0      0
#> Claim#51      0      0      0      0      0      0      0      0      0      0
#> Claim#52      0      0      0      0      0      0      0      0      0      0
#> Claim#60      0      0      0      0      0      0      0      0      0      0
#> Claim#68      0      0      0      0      0      0      0      0      0      0
#>          [,129] [,130]
#> Claim#10      0      0
#> Claim#13      0      0
#> Claim#16      0      0
#> Claim#22      0      0
#> Claim#32      0      0
#> Claim#37      0      0
#> Claim#41      0      0
#> Claim#43      0      0
#> Claim#44      0      0
#> Claim#45      0      0
#> Claim#49      0      0
#> Claim#51      0      0
#> Claim#52      0      0
#> Claim#60      0      0
#> Claim#68      0      0