Simplified version of sicr

This function offers a simplified version of sicr(). It can be used to used to get the gross results for certain development years and reserve classes without building all the for sicr() required objects in advance.

Usage

sicr_single(
  dev_year,
  reserve_class,
  last_orig_year,
  pools,
  indices,
  age_shift = NULL,
  mortality = NULL,
  n = 100
)

Arguments

dev_year

Integer or integer vector. Must be greater or equal to 1.

reserve_class

Integer or integer vector. Must be greater or equal to 0.

last_orig_year

Last origin year.

pools

List as output of generate_pools().

indices

Dataframe for indexation, see details of prepare_data().

age_shift

Dataframe, see description of age_shift_xmpl. Default: NULL
Only necessary if annuities shall be considered.

mortality

Dataframe, see description of mortality_xmpl. Default: NULL
Only necessary if annuities shall be considered.

n

Number of simulations. Default: 100

Value

Dataframe with columns 'Reserve_class', 'Dev_year' and 'Best_estimate'.

Examples

# this example uses data provided with this package
extended_claims_data <- prepare_data(claims_data = claims_data_xmpl,
                                     indices = indices_xmpl,
                                     threshold = 400000,
                                     first_orig_year = 1989,
                                     last_orig_year = 2023,
                                     expected_year_of_growing_large = 3,
                                     reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                                     pool_of_annuities = pool_of_annuities_xmpl)

pools <- generate_pools(extended_claims_data = extended_claims_data,
                        reserve_classes = c(1, 200001, 400001, 700001, 1400001),
                        years_for_pools = 2014:2023,
                        start_of_tail = 17,
                        end_of_tail = 50,
                        lower_outlier_limit = -Inf,
                        upper_outlier_limit = Inf,
                        pool_of_annuities = pool_of_annuities_xmpl)

sicr_single(dev_year = 1:5,
            reserve_class = 0:3,
            last_orig_year = 2023,
            pools = pools,
            indices = indices_xmpl,
            age_shift = age_shift_xmpl,
            mortality = mortality_xmpl)
#>    Reserve_class Dev_year Best_estimate
#> 1              0        5      496.4824
#> 2              0        4    29159.6982
#> 3              0        3    59002.6499
#> 4              0        2    26128.6079
#> 5              0        1    45161.6761
#> 6              1        5   115173.1039
#> 7              1        4   211239.8127
#> 8              1        3   319001.4628
#> 9              1        2   240684.4558
#> 10             1        1   288998.5410
#> 11             2        5   315293.1161
#> 12             2        4   273100.3709
#> 13             2        3   307705.3320
#> 14             2        2   332000.4126
#> 15             2        1   362691.5158
#> 16             3        5   337262.7987
#> 17             3        4   447972.9113
#> 18             3        3   394670.7270
#> 19             3        2   433412.2716
#> 20             3        1   482260.3085